feat(6748): extract service api ops configuration docs
