+++ crates/kamn-core/tests/service_api_ops_configuration_docs.rs
@@ -1,2111 +1,27 @@
 const DOC: &str = include_str!("../../../docs/ops/configuration.md");
 
-#[test]
-fn service_api_ops_configuration_contains_signer_secret_zeroization_controls() {
-    assert!(
-        DOC.contains("## Signer Secret Decode Buffer Zeroization Controls (Issues #4165, #4166)")
-    );
-    assert!(DOC.contains("signer_secret_source_precedence_zeroization_status=verified"));
-    assert!(DOC.contains("signer_private_key_parse_zeroization_status=verified"));
-    assert!(DOC.contains("signer_transient_key_material_zeroization_status=verified"));
-    assert!(DOC.contains("signer_secret_source_precedence_violation"));
-    assert!(DOC.contains("managed_signer_private_key_adapter_unsupported"));
-    assert!(DOC.contains(
-        "signer::tests::regression_signer_secret_source_precedence_failure_zeroizes_env_secret_buffer"
-    ));
-    assert!(DOC.contains(
-        "signer::tests::unit_build_kolme_live_managed_signing_key_zeroizes_transient_key_material"
-    ));
-    assert!(DOC.contains("Regression: #4165"));
-    assert!(DOC.contains("Regression: #4166"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_dependency_license_metadata_governance_remediation_markers(
-) {
-    assert!(DOC.contains(
-        "## Dependency-License Metadata Governance CI/Local Boundary Contract (Issue #4035)"
-    ));
-    assert!(DOC.contains(
-        "metadata_governance_reason_taxonomy_version=kamn.ci.dependency-license-metadata-governance-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "metadata_governance_reason_codes_csv=expected_license_empty,no_crate_manifests_found,license_policy_file_not_found,license_policy_marker_mismatch,manifest_not_found,manifest_invalid_toml,package_section_missing,license_missing,license_mismatch,metadata_governance_local_heavy_opt_in_required"
-    ));
-    assert!(DOC.contains("metadata_governance_remediation_map_version=v1"));
-    for reason_code in [
-        "expected_license_empty",
-        "no_crate_manifests_found",
-        "license_policy_file_not_found",
-        "license_policy_marker_mismatch",
-        "manifest_not_found",
-        "manifest_invalid_toml",
-        "package_section_missing",
-        "license_missing",
-        "license_mismatch",
-        "metadata_governance_local_heavy_opt_in_required",
-    ] {
-        assert!(
-            DOC.contains(&format!("metadata_governance_remediation.{reason_code}=")),
-            "missing ops dependency-license remediation marker for reason code {reason_code}"
-        );
+fn assert_doc_contains_all(markers: &[&str]) {
+    for marker in markers {
+        assert!(DOC.contains(marker), "missing doc marker: {marker}");
     }
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test ci_strategy_docs doc_enforces_dependency_license_metadata_remediation_markers_cover_reason_codes -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #4035"));
 }
 
-#[test]
-fn service_api_ops_configuration_contains_sbom_provenance_artifact_generator_markers() {
-    assert!(DOC.contains("## SBOM-Provenance Artifact Generator Contract (Issue #4036)"));
-    assert!(DOC.contains(
-        "sbom_provenance_lane_schema_version=kamn.runtime.sbom-provenance-artifact-report.v1"
-    ));
-    assert!(DOC.contains(
-        "sbom_provenance_artifact_schema_version=kamn.runtime.sbom-provenance-artifact-schema.v1"
-    ));
-    assert!(DOC.contains(
-        "sbom_provenance_fixture_schema_version=kamn.ci.sbom-provenance-artifact-fixture-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "sbom_provenance_reason_taxonomy_version=kamn.runtime.sbom-provenance-artifact-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "sbom_provenance_reason_codes_csv=sbom_provenance_profile_contract_violation,sbom_provenance_runtime_budget_exceeded"
-    ));
-    assert!(DOC.contains("sbom_schema_version=spdx-2.3"));
-    assert!(DOC.contains("provenance_schema_version=slsa-v1"));
-    assert!(DOC.contains("release_manifest_required_artifact_id=sbom_provenance"));
-    assert!(DOC.contains(
-        "cargo run -p kamn-core --bin sbom_provenance_artifact_generator_contract -- --profile baseline --mode dry-run --ci-fast-gate PASS --max-seconds 120 --output-json /tmp/sbom-provenance-baseline.json"
-    ));
-    assert!(DOC.contains(
-        "cargo run -p kamn-core --bin sbom_provenance_artifact_generator_contract -- --profile injected-drift --mode dry-run --ci-fast-gate PASS --max-seconds 120 --output-json /tmp/sbom-provenance-injected-drift.json"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test sbom_provenance_artifact_generator_contract -- --nocapture"
-    ));
-    assert!(DOC.contains("Regression: #4036"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_sbom_provenance_release_gonogo_checker_markers() {
-    assert!(DOC.contains("## SBOM-Provenance Release Go-No-Go Checker Contract (Issue #4037)"));
-    assert!(DOC.contains(
-        "sbom_provenance_release_gonogo_checker_schema_version=kamn.runtime.sbom-provenance-release-gonogo-checker-report.v1"
-    ));
-    assert!(DOC.contains(
-        "sbom_provenance_release_gonogo_checker_reason_taxonomy_version=kamn.runtime.sbom-provenance-release-gonogo-checker-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "sbom_provenance_release_gonogo_checker_reason_codes_csv=sbom_provenance_artifact_marker_missing,sbom_provenance_artifact_marker_invalid,sbom_provenance_artifact_decision_not_go,sbom_provenance_docs_parity_marker_missing,sbom_provenance_runtime_budget_exceeded"
-    ));
-    assert!(DOC.contains(
-        "sbom_provenance_release_gonogo_required_artifact_markers_csv=schema_version,artifact_schema_version,fixture_schema_version,reason_taxonomy_version,release_manifest_required_artifact_id,status,final_decision,reason_code"
-    ));
-    assert!(DOC.contains(
-        "sbom_provenance_release_gonogo_docs_parity_required_markers_csv=sbom_provenance_release_gonogo_checker_schema_version,sbom_provenance_release_gonogo_checker_reason_taxonomy_version,sbom_provenance_release_gonogo_checker_reason_codes_csv,sbom_provenance_release_gonogo_required_artifact_markers_csv"
-    ));
-    assert!(DOC.contains(
-        "python3 scripts/deploy/sbom_provenance_release_gonogo_checker_contract.py --artifact-json /tmp/sbom-provenance-baseline.json --ci-strategy-doc docs/ci/strategy.md --ops-configuration-doc docs/ops/configuration.md --max-seconds 120 --output-json /tmp/sbom-provenance-release-gonogo-checker.json"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test sbom_provenance_release_gonogo_checker_contract -- --nocapture"
-    ));
-    assert!(DOC.contains("Regression: #4037"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_async_backpressure_failure_modes() {
-    assert!(DOC.contains("## Async API Backpressure Failure Modes (Issue #4315)"));
-    assert!(DOC.contains(
-        "service_api_backpressure_reason_taxonomy_version=kamn.runtime.service-api.lifecycle-rejection-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains("service_api_ingress_concurrency_limit_exceeded"));
-    assert!(DOC.contains("service_api_ingress_rate_limit_exceeded"));
-    assert!(DOC.contains("service_api_ingress_sender_rate_limit_exceeded"));
-    assert!(DOC.contains("admission_inflight_budget_status=verified"));
-    assert!(DOC.contains("admission_queue_budget_status=verified"));
-    assert!(DOC.contains("admission_inflight_budget_limit=32"));
-    assert!(DOC.contains("admission_queue_budget_limit=1"));
-    assert!(DOC.contains(
-        "admission_budget_reason_taxonomy_version=kamn.runtime.service-api-admission-budget-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "admission_budget_reason_codes_csv=admission_inflight_budget_mismatch,admission_queue_budget_mismatch"
-    ));
-    assert!(DOC.contains("admission_decision_taxonomy_status=verified"));
-    assert!(DOC.contains("admission_decision_accept_status=verified"));
-    assert!(DOC.contains("admission_decision_defer_status=verified"));
-    assert!(DOC.contains("admission_decision_reject_status=verified"));
-    assert!(DOC.contains(
-        "admission_decision_reason_taxonomy_version=kamn.runtime.service-api-admission-decision-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "admission_decision_reason_codes_csv=admission_decision_accept,admission_decision_defer,admission_decision_reject"
-    ));
-    assert!(DOC.contains("service_api_axum_policy_admission_inflight_budget_limit_mismatch"));
-    assert!(DOC.contains("service_api_axum_policy_admission_queue_budget_limit_mismatch"));
-    assert!(
-        DOC.contains("service_api_axum_policy_admission_decision_reason_taxonomy_version_mismatch")
-    );
-    assert!(DOC.contains("service_api_axum_policy_admission_decision_reason_codes_csv_mismatch"));
-    assert!(DOC.contains("fail-closed response contract"));
-    assert!(DOC.contains("Regression: #4315"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_tenant_isolation_matrix_markers() {
-    assert!(DOC.contains("## Service API Tenant-Isolation Matrix Contract (Issue #4058)"));
-    assert!(DOC.contains(
-        "service_api_tenant_isolation_matrix_reason_taxonomy_version=kamn.runtime.service-api-tenant-isolation-matrix-policy-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "service_api_tenant_isolation_matrix_reason_codes_csv=ci_fast_gate_failed,service_api_tenant_isolation_policy_schema_mismatch,service_api_tenant_isolation_policy_status_invalid,service_api_tenant_isolation_policy_final_decision_invalid,service_api_tenant_isolation_policy_final_decision_mismatch,service_api_tenant_isolation_policy_lane_mode_invalid,service_api_tenant_isolation_policy_matrix_schema_mismatch,service_api_tenant_isolation_policy_matrix_rows_invalid,service_api_tenant_isolation_policy_matrix_row_count_mismatch,service_api_tenant_isolation_policy_matrix_row_duplicate,service_api_tenant_isolation_policy_matrix_row_id_invalid,service_api_tenant_isolation_policy_matrix_row_missing,service_api_tenant_isolation_policy_matrix_row_status_mismatch,service_api_tenant_isolation_policy_matrix_row_leakage_result_mismatch,service_api_tenant_isolation_policy_matrix_row_reason_code_mismatch,service_api_tenant_isolation_policy_matrix_row_selector_mismatch,service_api_tenant_isolation_policy_marker_missing,service_api_tenant_isolation_policy_execution_reason_code_mismatch,service_api_tenant_isolation_policy_command_count_invalid,service_api_tenant_isolation_policy_command_count_mismatch,service_api_tenant_isolation_policy_elapsed_seconds_invalid,service_api_tenant_isolation_policy_max_seconds_invalid,service_api_tenant_isolation_policy_runtime_budget_exceeded,service_api_tenant_isolation_policy_docs_marker_missing"
-    ));
-    assert!(DOC.contains(
-        "service_api_tenant_isolation_matrix_matrix_schema_version=kamn.runtime.service-api-tenant-isolation-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "service_api_tenant_isolation_matrix_required_row_ids_csv=m2_abac_cross_tenant_visibility_denied,m8_cross_owner_retention_and_shred_denied,m9_cross_owner_dispatch_and_presence_denied,m9_gateway_cross_owner_presence_denied"
-    ));
-    assert!(DOC.contains("m2_abac_scope_denied"));
-    assert!(DOC.contains("m8_compliance_owner_scope_denied"));
-    assert!(DOC.contains("m9_realtime_owner_scope_denied"));
-    assert!(DOC.contains("service_api_tenant_isolation_policy_matrix_row_status_mismatch"));
-    assert!(DOC.contains(
-        "bash scripts/runtime/validate_service_api_tenant_isolation_matrix_live.sh --mode dry-run --output-json /tmp/service-api-tenant-isolation-matrix-live-summary.json"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/check_service_api_tenant_isolation_matrix_live_policy.sh --report-file /tmp/service-api-tenant-isolation-matrix-live-summary.json --expected-final-decision GO --ci-fast-gate PASS --output-json /tmp/service-api-tenant-isolation-matrix-policy.json"
-    ));
-    assert!(DOC.contains("Regression: #4058"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_api_version_policy_markers() {
-    assert!(DOC.contains("## API Version-Policy Contract (Issue #4041)"));
-    assert!(DOC.contains(
-        "api_version_policy_reason_taxonomy_version=kamn.runtime.api-version-policy-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "api_version_policy_reason_codes_csv=ci_fast_gate_failed,api_version_policy_schema_mismatch,api_version_policy_status_invalid,api_version_policy_final_decision_invalid,api_version_policy_final_decision_mismatch,api_version_policy_lane_mode_invalid,api_version_policy_fixture_schema_mismatch,api_version_policy_fixture_rows_invalid,api_version_policy_fixture_row_count_mismatch,api_version_policy_fixture_row_duplicate,api_version_policy_fixture_row_id_invalid,api_version_policy_fixture_row_missing,api_version_policy_fixture_row_status_mismatch,api_version_policy_fixture_row_decision_mismatch,api_version_policy_fixture_row_reason_code_mismatch,api_version_policy_fixture_row_version_mismatch,api_version_policy_fixture_row_window_mismatch,api_version_policy_marker_missing,api_version_policy_execution_reason_code_mismatch,api_version_policy_command_count_invalid,api_version_policy_command_count_mismatch,api_version_policy_elapsed_seconds_invalid,api_version_policy_max_seconds_invalid,api_version_policy_runtime_budget_exceeded,api_version_policy_docs_marker_missing"
-    ));
-    assert!(DOC.contains(
-        "api_version_policy_fixture_schema_version=kamn.runtime.api-version-policy-fixture-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "api_version_policy_fixture_path=fixtures/runtime/api_version_policy_fixture_matrix.txt"
-    ));
-    assert!(DOC.contains(
-        "api_version_policy_required_row_ids_csv=v1_messages_send,v2_channels_create,v0_messages_send,v3_future_route"
-    ));
-    assert!(DOC.contains("api_version_unsupported_window"));
-    assert!(DOC.contains("api_version_policy_fixture_row_status_mismatch"));
-    assert!(DOC.contains(
-        "bash scripts/runtime/validate_api_version_policy_live.sh --mode dry-run --output-json /tmp/api-version-policy-live-summary.json"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/check_api_version_policy_live_policy.sh --report-file /tmp/api-version-policy-live-summary.json --expected-final-decision GO --ci-fast-gate PASS --output-json /tmp/api-version-policy-live-policy.json"
-    ));
-    assert!(DOC.contains("Regression: #4041"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_request_response_schema_compatibility_markers() {
-    assert!(DOC.contains("## Request-Response Schema Compatibility Contract (Issue #4042)"));
-    assert!(DOC.contains(
-        "request_response_schema_compatibility_reason_taxonomy_version=kamn.runtime.request-response-schema-compatibility-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "request_response_schema_compatibility_reason_codes_csv=ci_fast_gate_failed,request_response_schema_compatibility_schema_mismatch,request_response_schema_compatibility_status_invalid,request_response_schema_compatibility_final_decision_invalid,request_response_schema_compatibility_final_decision_mismatch,request_response_schema_compatibility_lane_mode_invalid,request_response_schema_compatibility_fixture_schema_mismatch,request_response_schema_compatibility_fixture_rows_invalid,request_response_schema_compatibility_fixture_row_count_mismatch,request_response_schema_compatibility_fixture_row_duplicate,request_response_schema_compatibility_fixture_row_id_invalid,request_response_schema_compatibility_fixture_row_missing,request_response_schema_compatibility_fixture_row_status_mismatch,request_response_schema_compatibility_fixture_row_decision_mismatch,request_response_schema_compatibility_fixture_row_reason_code_mismatch,request_response_schema_compatibility_fixture_row_version_pair_mismatch,request_response_schema_compatibility_fixture_row_change_class_mismatch,request_response_schema_compatibility_marker_missing,request_response_schema_compatibility_execution_reason_code_mismatch,request_response_schema_compatibility_command_count_invalid,request_response_schema_compatibility_command_count_mismatch,request_response_schema_compatibility_elapsed_seconds_invalid,request_response_schema_compatibility_max_seconds_invalid,request_response_schema_compatibility_runtime_budget_exceeded,request_response_schema_compatibility_docs_marker_missing"
-    ));
-    assert!(DOC.contains(
-        "request_response_schema_compatibility_fixture_schema_version=kamn.runtime.request-response-schema-compatibility-fixture-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "request_response_schema_compatibility_fixture_path=fixtures/runtime/request_response_schema_compatibility_fixture_matrix.txt"
-    ));
-    assert!(DOC.contains(
-        "request_response_schema_compatibility_required_row_ids_csv=v1_to_v2_messages_send_optional_request_addition,v1_to_v2_channels_create_optional_response_addition,v1_to_v2_messages_get_required_response_removal,v1_to_v2_tasks_create_required_request_removal"
-    ));
-    assert!(DOC.contains("schema_pair_breaking_change_detected"));
-    assert!(DOC.contains("request_response_schema_compatibility_fixture_row_status_mismatch"));
-    assert!(DOC.contains(
-        "bash scripts/runtime/validate_request_response_schema_compatibility_live.sh --mode dry-run --output-json /tmp/request-response-schema-compatibility-live-summary.json"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/check_request_response_schema_compatibility_live_policy.sh --report-file /tmp/request-response-schema-compatibility-live-summary.json --expected-final-decision GO --ci-fast-gate PASS --output-json /tmp/request-response-schema-compatibility-live-policy.json"
-    ));
-    assert!(DOC.contains("Regression: #4042"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_api_compatibility_matrix_local_heavy_markers() {
-    assert!(DOC.contains("## API Compatibility Matrix Local-Heavy Contract (Issue #4043)"));
-    assert!(DOC.contains(
-        "api_compatibility_matrix_local_heavy_reason_taxonomy_version=kamn.runtime.api-compatibility-matrix-local-heavy-policy-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "api_compatibility_matrix_local_heavy_reason_codes_csv=ci_fast_gate_failed,api_compatibility_matrix_local_heavy_policy_schema_mismatch,api_compatibility_matrix_local_heavy_policy_status_invalid,api_compatibility_matrix_local_heavy_policy_final_decision_invalid,api_compatibility_matrix_local_heavy_policy_final_decision_mismatch,api_compatibility_matrix_local_heavy_policy_lane_mode_invalid,api_compatibility_matrix_local_heavy_policy_artifact_schema_mismatch,api_compatibility_matrix_local_heavy_policy_fixture_schema_mismatch,api_compatibility_matrix_local_heavy_policy_fixture_rows_invalid,api_compatibility_matrix_local_heavy_policy_fixture_row_count_mismatch,api_compatibility_matrix_local_heavy_policy_fixture_row_duplicate,api_compatibility_matrix_local_heavy_policy_fixture_row_id_invalid,api_compatibility_matrix_local_heavy_policy_fixture_row_missing,api_compatibility_matrix_local_heavy_policy_fixture_row_status_mismatch,api_compatibility_matrix_local_heavy_policy_fixture_row_decision_mismatch,api_compatibility_matrix_local_heavy_policy_fixture_row_reason_code_mismatch,api_compatibility_matrix_local_heavy_policy_fixture_row_version_pair_mismatch,api_compatibility_matrix_local_heavy_policy_fixture_row_route_selector_mismatch,api_compatibility_matrix_local_heavy_policy_fixture_row_change_class_mismatch,api_compatibility_matrix_local_heavy_policy_marker_missing,api_compatibility_matrix_local_heavy_policy_execution_reason_code_mismatch,api_compatibility_matrix_local_heavy_policy_command_count_invalid,api_compatibility_matrix_local_heavy_policy_command_count_mismatch,api_compatibility_matrix_local_heavy_policy_elapsed_seconds_invalid,api_compatibility_matrix_local_heavy_policy_max_seconds_invalid,api_compatibility_matrix_local_heavy_policy_runtime_budget_exceeded,api_compatibility_matrix_local_heavy_policy_local_heavy_opt_in_required,api_compatibility_matrix_local_heavy_policy_local_heavy_scope_mismatch,api_compatibility_matrix_local_heavy_policy_docs_marker_missing"
-    ));
-    assert!(DOC.contains(
-        "api_compatibility_matrix_local_heavy_fixture_schema_version=kamn.runtime.api-compatibility-matrix-local-heavy-fixture-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "api_compatibility_matrix_local_heavy_fixture_path=fixtures/runtime/api_compatibility_matrix_local_heavy_fixture_matrix.txt"
-    ));
-    assert!(DOC.contains(
-        "api_compatibility_matrix_local_heavy_required_row_ids_csv=v1_to_v2_messages_send_optional_request_addition,v1_to_v2_channels_create_optional_response_addition,v1_to_v2_tasks_create_required_request_removal,v1_to_v2_messages_get_required_response_removal,v1_to_v2_messages_send_enum_variant_removal"
-    ));
-    assert!(DOC.contains("incompatible_request_breaking_change"));
-    assert!(DOC.contains("incompatible_response_breaking_change"));
-    assert!(DOC.contains("incompatible_enum_breaking_change"));
-    assert!(DOC.contains("api_compatibility_matrix_local_heavy_policy_fixture_row_status_mismatch"));
-    assert!(DOC.contains(
-        "bash scripts/runtime/validate_api_compatibility_matrix_local_heavy_live.sh --mode dry-run --ci-fast-gate PASS --output-json /tmp/api-compatibility-matrix-local-heavy-summary.json"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/check_api_compatibility_matrix_local_heavy_live_policy.sh --report-file /tmp/api-compatibility-matrix-local-heavy-summary.json --expected-final-decision GO --ci-fast-gate PASS --output-json /tmp/api-compatibility-matrix-local-heavy-policy.json"
-    ));
-    assert!(DOC.contains("Regression: #4043"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_sqlite_crash_recovery_ci_dry_run_governance_markers() {
-    assert!(DOC.contains(
-        "## SQLite Crash-Recovery CI Dry-Run Durability Governance Contract (Issue #4014)"
-    ));
-    assert!(DOC.contains(
-        "sqlite_crash_recovery_ci_dry_run_reason_taxonomy_version=kamn.ci.sqlite-crash-recovery-ci-dry-run-governance-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "sqlite_crash_recovery_ci_dry_run_reason_codes_csv=sqlite_crash_recovery_ci_dry_run_argument_invalid,sqlite_crash_recovery_ci_dry_run_threshold_contract_violation,sqlite_crash_recovery_ci_dry_run_report_contract_violation,sqlite_crash_recovery_ci_dry_run_runtime_budget_exceeded,sqlite_crash_recovery_ci_dry_run_fast_mode_selector_drift,sqlite_crash_recovery_ci_dry_run_workflow_exclusion_drift,sqlite_crash_recovery_ci_dry_run_docs_marker_parity_drift,sqlite_crash_recovery_ci_dry_run_docs_remediation_marker_missing"
-    ));
-    assert!(DOC.contains(
-        "sqlite_crash_recovery_ci_dry_run_threshold_fixture_path=fixtures/ci/sqlite_crash_recovery_ci_dry_run_governance_thresholds.env"
-    ));
-    assert!(DOC.contains("sqlite_crash_recovery_ci_dry_run_max_seconds=120"));
-    assert!(DOC.contains(
-        "sqlite_crash_recovery_ci_dry_run_fast_mode_required_entry=cargo test -p kamn-core --test sqlite_crash_recovery_ci_dry_run_governance_contract -- --nocapture"
-    ));
-    assert!(DOC.contains(
-        "sqlite_crash_recovery_ci_dry_run_fast_mode_forbidden_entry=bash \"$ROOT_DIR/scripts/runtime/validate_sqlite_crash_recovery_live.sh\" --mode run"
-    ));
-    assert!(DOC.contains(
-        "sqlite_crash_recovery_ci_dry_run_workflow_forbidden_entry=bash scripts/runtime/validate_sqlite_crash_recovery_live.sh --mode run"
-    ));
-    assert!(DOC.contains(
-        "python3 scripts/ci/check_sqlite_crash_recovery_ci_dry_run_governance.py --sqlite-crash-recovery-summary-report-file /tmp/sqlite-crash-recovery-live-summary.json --sqlite-crash-recovery-policy-report-file /tmp/sqlite-crash-recovery-live-policy.json --sqlite-crash-recovery-contract-lane-report-file /tmp/sqlite-crash-recovery-live-contract-lane-report.json --threshold-file fixtures/ci/sqlite_crash_recovery_ci_dry_run_governance_thresholds.env --strategy-doc docs/ci/strategy.md --ops-doc docs/ops/configuration.md --workflow-file .github/workflows/ci-fast-gate.yml --ci-tools-file scripts/ci/test_ci_tools.sh --output-json /tmp/sqlite-crash-recovery-ci-dry-run-governance-report.json"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test sqlite_crash_recovery_ci_dry_run_governance_contract -- --nocapture"
-    ));
-    assert!(DOC.contains("sqlite_crash_recovery_ci_dry_run_remediation_map_version=v1"));
-    for reason_code in [
-        "sqlite_crash_recovery_ci_dry_run_argument_invalid",
-        "sqlite_crash_recovery_ci_dry_run_threshold_contract_violation",
-        "sqlite_crash_recovery_ci_dry_run_report_contract_violation",
-        "sqlite_crash_recovery_ci_dry_run_runtime_budget_exceeded",
-        "sqlite_crash_recovery_ci_dry_run_fast_mode_selector_drift",
-        "sqlite_crash_recovery_ci_dry_run_workflow_exclusion_drift",
-        "sqlite_crash_recovery_ci_dry_run_docs_marker_parity_drift",
-        "sqlite_crash_recovery_ci_dry_run_docs_remediation_marker_missing",
-    ] {
-        assert!(
-            DOC.contains(&format!(
-                "sqlite_crash_recovery_ci_dry_run_remediation.{reason_code}="
-            )),
-            "missing ops remediation marker for reason code {reason_code}"
-        );
+fn assert_doc_contains_prefixed_entries(prefix: &str, codes: &[&str]) {
+    for code in codes {
+        let marker = format!("{prefix}.{code}=");
+        assert!(DOC.contains(&marker), "missing doc marker: {marker}");
     }
-    assert!(DOC.contains("Regression: #4014"));
 }
 
-#[test]
-fn service_api_ops_configuration_contains_journal_wal_partial_write_fault_injection_markers() {
-    assert!(DOC.contains("## Journal/WAL Partial-Write Fault Injection Contracts (Issue #4016)"));
-    assert!(DOC.contains(
-        "journal_wal_partial_write_fixture_path=fixtures/runtime/journal_wal_partial_write_fault_matrix.txt"
-    ));
-    assert!(DOC.contains(
-        "journal_wal_partial_write_fixture_schema_version=kamn.runtime.journal-wal-partial-write-fault-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "journal_wal_partial_write_reason_taxonomy_version=kamn.runtime.journal-wal-partial-write-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "journal_wal_partial_write_reason_codes_csv=partial_snapshot_file_write_recovered_from_journal,partial_journal_tail_write_fail_closed,partial_snapshot_without_journal_repaired"
-    ));
-    assert!(DOC.contains(
-        "journal_wal_partial_write_required_fault_modes_csv=partial_snapshot_file_write,partial_journal_tail_write,partial_snapshot_without_journal"
-    ));
-    assert!(DOC.contains("partial_snapshot_file_write -> recovery_clean"));
-    assert!(DOC.contains("partial_journal_tail_write -> fail_closed_corrupt_tail"));
-    assert!(DOC.contains("partial_snapshot_without_journal -> recovery_repaired_corrupt_payload"));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test journal_wal_partial_write_fault_contract -- --nocapture"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_journal_wal_partial_write_fault_injection_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #4016"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_sqlite_crash_restart_local_heavy_lane_markers() {
-    assert!(
-        DOC.contains("## SQLite Crash-Restart Local-Heavy Lane Artifact Contract (Issue #4017)")
-    );
-    assert!(DOC.contains(
-        "sqlite_crash_restart_local_heavy_lane_schema_version=kamn.runtime.sqlite-crash-restart-local-heavy-lane-report.v1"
-    ));
-    assert!(DOC.contains(
-        "sqlite_crash_restart_local_heavy_artifact_schema_version=kamn.runtime.sqlite-crash-restart-local-heavy-artifact-schema.v1"
-    ));
-    assert!(DOC.contains(
-        "sqlite_crash_restart_local_heavy_reason_taxonomy_version=kamn.runtime.sqlite-crash-restart-local-heavy-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "sqlite_crash_restart_local_heavy_reason_codes_csv=crash_restart_profile_restart_status_mismatch,crash_restart_profile_corruption_status_mismatch,crash_restart_profile_combined_status_mismatch"
-    ));
-    assert!(DOC.contains(
-        "sqlite_crash_restart_local_heavy_required_profiles_csv=restart,corruption,combined"
-    ));
-    assert!(DOC.contains("profile=restart -> restart_drill_status=verified"));
-    assert!(DOC.contains("profile=corruption -> corruption_drill_status=verified"));
-    assert!(DOC.contains(
-        "profile=combined -> restart_drill_status=verified + corruption_drill_status=verified"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/run_sqlite_crash_restart_local_heavy_lane.sh --profile combined --mode dry-run --ci-fast-gate PASS --max-seconds 240 --output-json /tmp/sqlite-crash-restart-local-heavy-lane-report.json"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test sqlite_crash_restart_local_heavy_lane_contract -- --nocapture"
-    ));
-    assert!(DOC.contains("Regression: #4017"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_local_heavy_capacity_load_lane_markers() {
-    assert!(DOC.contains("## Local-Heavy Capacity/Load Lane Artifact Contract (Issue #4004)"));
-    assert!(DOC.contains(
-        "local_heavy_capacity_load_lane_schema_version=kamn.runtime.local-heavy-capacity-load-lane-report.v1"
-    ));
-    assert!(DOC.contains(
-        "local_heavy_capacity_load_artifact_schema_version=kamn.runtime.local-heavy-capacity-load-artifact-schema.v1"
-    ));
-    assert!(DOC.contains(
-        "local_heavy_capacity_load_reason_taxonomy_version=kamn.runtime.local-heavy-capacity-load-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "local_heavy_capacity_load_reason_codes_csv=local_heavy_capacity_load_profile_threshold_breach,local_heavy_capacity_load_runtime_budget_exceeded"
-    ));
-    assert!(DOC.contains("local_heavy_capacity_load_required_profiles_csv=baseline,fault"));
-    assert!(DOC
-        .contains("profile=baseline -> status=pass + final_decision=GO + profile_status=verified"));
-    assert!(
-        DOC.contains("profile=fault -> status=fail + final_decision=NO-GO + profile_status=failed")
-    );
-    assert!(DOC.contains(
-        "bash scripts/runtime/run_local_heavy_capacity_load_lane.sh --profile baseline --mode dry-run --max-seconds 120 --output-json /tmp/local-heavy-capacity-load-baseline.json"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/run_local_heavy_capacity_load_lane.sh --profile fault --mode dry-run --max-seconds 120 --output-json /tmp/local-heavy-capacity-load-fault.json"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test local_heavy_capacity_load_lane_contract -- --nocapture"
-    ));
-    assert!(DOC.contains("Regression: #4004"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_local_heavy_redaction_validation_lane_markers() {
-    assert!(
-        DOC.contains("## Local-Heavy Redaction Validation Lane Artifact Contract (Issue #4079)")
-    );
-    assert!(DOC.contains(
-        "local_heavy_redaction_validation_lane_schema_version=kamn.runtime.local-heavy-redaction-validation-lane-report.v1"
-    ));
-    assert!(DOC.contains(
-        "local_heavy_redaction_validation_artifact_schema_version=kamn.runtime.local-heavy-redaction-validation-artifact-schema.v1"
-    ));
-    assert!(DOC.contains(
-        "local_heavy_redaction_validation_reason_taxonomy_version=kamn.runtime.local-heavy-redaction-validation-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "local_heavy_redaction_validation_reason_codes_csv=local_heavy_redaction_sensitive_pattern_detected,local_heavy_redaction_runtime_budget_exceeded"
-    ));
-    assert!(DOC
-        .contains("local_heavy_redaction_validation_required_profiles_csv=baseline,injected-leak"));
-    assert!(DOC.contains(
-        "profile=baseline -> status=pass + final_decision=GO + profile_status=verified + leak_marker_status=clear"
-    ));
-    assert!(DOC.contains(
-        "profile=injected-leak -> status=fail + final_decision=NO-GO + profile_status=failed + leak_marker_status=detected"
-    ));
-    assert!(
-        DOC.contains("profile=baseline -> leak_detection_count=0 + leaked_pattern_ids_csv=none")
-    );
-    assert!(DOC.contains(
-        "profile=injected-leak -> leak_detection_count=3 + leaked_pattern_ids_csv=raw_signer_secret,pii_email,pii_phone"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/run_local_heavy_redaction_validation_lane.sh --profile baseline --mode dry-run --max-seconds 120 --output-json /tmp/local-heavy-redaction-validation-baseline.json"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/run_local_heavy_redaction_validation_lane.sh --profile injected-leak --mode dry-run --max-seconds 120 --output-json /tmp/local-heavy-redaction-validation-injected-leak.json"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test local_heavy_redaction_validation_lane_contract -- --nocapture"
-    ));
-    assert!(DOC.contains("Regression: #4079"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_dependency_local_heavy_deep_scan_lane_markers() {
-    assert!(
-        DOC.contains("## Dependency Local-Heavy Deep Scan Lane Artifact Contract (Issue #4032)")
-    );
-    assert!(DOC.contains(
-        "dependency_local_heavy_deep_scan_lane_schema_version=kamn.runtime.dependency-local-heavy-deep-scan-lane-report.v1"
-    ));
-    assert!(DOC.contains(
-        "dependency_local_heavy_deep_scan_artifact_schema_version=kamn.runtime.dependency-local-heavy-deep-scan-artifact-schema.v1"
-    ));
-    assert!(DOC.contains(
-        "dependency_local_heavy_deep_scan_fixture_schema_version=kamn.ci.dependency-local-heavy-deep-scan-fixture-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "dependency_local_heavy_deep_scan_reason_taxonomy_version=kamn.runtime.dependency-local-heavy-deep-scan-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "dependency_local_heavy_deep_scan_reason_codes_csv=dependency_local_heavy_deep_scan_profile_threshold_exceeded,dependency_local_heavy_deep_scan_runtime_budget_exceeded"
-    ));
-    assert!(DOC
-        .contains("dependency_local_heavy_deep_scan_required_profiles_csv=baseline,injected-risk"));
-    assert!(DOC.contains(
-        "profile=baseline -> status=pass + final_decision=GO + profile_status=verified + critical_count=0 + high_count=0"
-    ));
-    assert!(DOC.contains(
-        "profile=injected-risk -> status=fail + final_decision=NO-GO + profile_status=failed + critical_count=1 + high_count=2"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/run_dependency_local_heavy_deep_scan_lane.sh --profile baseline --mode dry-run --max-seconds 180 --output-json /tmp/dependency-local-heavy-deep-scan-baseline.json"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/run_dependency_local_heavy_deep_scan_lane.sh --profile injected-risk --mode dry-run --max-seconds 180 --output-json /tmp/dependency-local-heavy-deep-scan-injected-risk.json"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test dependency_local_heavy_deep_scan_lane_contract -- --nocapture"
-    ));
-    assert!(DOC.contains("Regression: #4032"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_lifecycle_artifact_integrity_markers() {
-    assert!(DOC.contains("## Tamper-Evident Lifecycle Artifact Integrity Contract (Issue #4081)"));
-    assert!(DOC.contains(
-        "lifecycle_artifact_integrity_schema_version=kamn.runtime.lifecycle-artifact-integrity-evidence.v1"
-    ));
-    assert!(DOC.contains(
-        "lifecycle_artifact_integrity_artifact_schema_version=kamn.runtime.lifecycle-artifact-integrity-schema.v1"
-    ));
-    assert!(DOC.contains(
-        "lifecycle_artifact_integrity_reason_taxonomy_version=kamn.runtime.lifecycle-artifact-integrity-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "lifecycle_artifact_integrity_reason_codes_csv=lifecycle_artifact_required_field_missing,lifecycle_artifact_marker_mismatch,lifecycle_artifact_hash_mismatch,lifecycle_artifact_reason_taxonomy_mismatch,lifecycle_artifact_reason_codes_csv_mismatch,lifecycle_artifact_expected_decision_mismatch"
-    ));
-    assert!(DOC.contains(
-        "lifecycle_artifact_integrity_hash_fields_csv=payload_hash_sha256,integrity_hash_sha256,provenance_hash_sha256"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/generate_lifecycle_artifact_integrity_evidence_bundle.sh --output-file /tmp/lifecycle-artifact-integrity-baseline.json --artifact-id lifecycle-artifact-baseline --lifecycle-stage retention --profile baseline --record-count 42 --ci-fast-gate PASS"
-    ));
-    assert!(DOC.contains(
-        "bash scripts/runtime/check_lifecycle_artifact_integrity_evidence_bundle.sh --bundle-file /tmp/lifecycle-artifact-integrity-baseline.json --expected-final-decision GO"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test lifecycle_artifact_integrity_contract -- --nocapture"
-    ));
-    assert!(DOC.contains("Regression: #4081"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_lifecycle_ci_dry_run_governance_markers() {
-    assert!(DOC.contains("## Lifecycle Artifact CI Dry-Run Governance Contract (Issue #4082)"));
-    assert!(DOC.contains(
-        "lifecycle_ci_dry_run_reason_taxonomy_version=kamn.ci.lifecycle-ci-dry-run-governance-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "lifecycle_ci_dry_run_reason_codes_csv=lifecycle_ci_dry_run_argument_invalid,lifecycle_ci_dry_run_threshold_contract_violation,lifecycle_ci_dry_run_report_contract_violation,lifecycle_ci_dry_run_lifecycle_marker_parity_drift,lifecycle_ci_dry_run_go_no_go_marker_parity_drift,lifecycle_ci_dry_run_runtime_budget_exceeded,lifecycle_ci_dry_run_fast_mode_selector_drift,lifecycle_ci_dry_run_workflow_exclusion_drift,lifecycle_ci_dry_run_docs_marker_parity_drift,lifecycle_ci_dry_run_docs_remediation_marker_missing"
-    ));
-    assert!(DOC.contains(
-        "lifecycle_ci_dry_run_threshold_fixture_path=fixtures/ci/lifecycle_ci_dry_run_governance_thresholds.env"
-    ));
-    assert!(DOC.contains("lifecycle_ci_dry_run_max_seconds=120"));
-    assert!(DOC.contains(
-        "lifecycle_ci_dry_run_fast_mode_required_entry=cargo test -p kamn-core --test lifecycle_ci_dry_run_governance_contract -- --nocapture"
-    ));
-    assert!(DOC.contains(
-        "lifecycle_ci_dry_run_fast_mode_forbidden_entry=bash \"$ROOT_DIR/scripts/runtime/run_go_no_go_gate_lane.sh\" --mode run"
-    ));
-    assert!(DOC.contains(
-        "lifecycle_ci_dry_run_workflow_forbidden_entry=bash scripts/runtime/run_go_no_go_gate_lane.sh --mode run"
-    ));
-    assert!(DOC.contains(
-        "python3 scripts/ci/check_lifecycle_ci_dry_run_governance.py --lifecycle-artifact-bundle-file /tmp/lifecycle-artifact-integrity-baseline.json --go-no-go-gate-report-file /tmp/go-no-go-gate-report.json --threshold-file fixtures/ci/lifecycle_ci_dry_run_governance_thresholds.env --strategy-doc docs/ci/strategy.md --ops-doc docs/ops/configuration.md --workflow-file .github/workflows/ci-fast-gate.yml --ci-tools-file scripts/ci/test_ci_tools.sh --output-json /tmp/lifecycle-ci-dry-run-governance-report.json"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test lifecycle_ci_dry_run_governance_contract -- --nocapture"
-    ));
-    assert!(DOC.contains("lifecycle_ci_dry_run_remediation_map_version=v1"));
-    for reason_code in [
-        "lifecycle_ci_dry_run_argument_invalid",
-        "lifecycle_ci_dry_run_threshold_contract_violation",
-        "lifecycle_ci_dry_run_report_contract_violation",
-        "lifecycle_ci_dry_run_lifecycle_marker_parity_drift",
-        "lifecycle_ci_dry_run_go_no_go_marker_parity_drift",
-        "lifecycle_ci_dry_run_runtime_budget_exceeded",
-        "lifecycle_ci_dry_run_fast_mode_selector_drift",
-        "lifecycle_ci_dry_run_workflow_exclusion_drift",
-        "lifecycle_ci_dry_run_docs_marker_parity_drift",
-        "lifecycle_ci_dry_run_docs_remediation_marker_missing",
-    ] {
-        assert!(
-            DOC.contains(&format!("lifecycle_ci_dry_run_remediation.{reason_code}=")),
-            "missing ops remediation marker for reason code {reason_code}"
-        );
-    }
-    assert!(DOC.contains("Regression: #4082"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_realtime_presence_mode_and_guardrail_markers() {
-    assert!(DOC.contains(
-        "## Realtime Presence Mode Gateway and Guardrail Contracts (Issues #5279, #5281, #5283)"
-    ));
-    assert!(DOC.contains("service_api_ws_presence_mode_status=verified"));
-    assert!(DOC.contains("service_api_ws_events_mode_header=x-kamn-events-mode"));
-    assert!(DOC.contains("service_api_ws_events_mode_presence_value=presence"));
-    assert!(DOC.contains(
-        "service_api_ws_presence_required_headers_csv=x-kamn-presence-owner-did,x-kamn-presence-target-agent-did,x-kamn-requester-agent-did"
-    ));
-    assert!(DOC.contains(
-        "service_api_ws_presence_optional_headers_csv=x-kamn-presence-target-owner-did,x-kamn-presence-gateway-node,x-kamn-presence-connected-since,x-kamn-presence-last-heartbeat,x-kamn-presence-capabilities"
-    ));
-    assert!(DOC.contains(
-        "service_api_ws_presence_fail_closed_reason_codes_csv=service_api_ws_events_mode_invalid,service_api_ws_presence_owner_did_header_missing,service_api_ws_presence_target_agent_did_header_missing,service_api_ws_presence_requester_agent_did_header_missing,m9_realtime_owner_scope_denied,m9_realtime_presence_visibility_denied"
-    ));
-    assert!(DOC.contains("service_api_ws_presence_event_type=m9.presence.snapshot"));
-    assert!(DOC.contains("service_api_ws_presence_transport_profile=websocket"));
-    assert!(DOC.contains("realtime_guardrail_burst_validation_status=verified"));
-    assert!(DOC.contains("replay_duplicate_reason_ordering_status=verified"));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node integration_service_api_endpoint_sender_anti_spam_burst_rounds_remain_deterministic -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node integration_service_api_endpoint_concurrency_rejection_reason_stays_stable_under_bounded_bursts -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node regression_service_api_endpoint_replay_duplicate_sequence_reason_ordering_stays_stable -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5283"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_phase6_archival_retry_policy_markers() {
-    assert!(
-        DOC.contains("## Phase-6 Archival Failure-Retry Policy Contracts (Issues #5285, #5287)")
-    );
-    assert!(DOC.contains("archival_retry_policy_status=verified"));
-    assert!(DOC.contains(
-        "archival_retry_reason_taxonomy_version=kamn.runtime.data-layer-m10-archival-retry-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "archival_retry_reason_codes_csv=m10_archival_retry_scheduled,m10_archival_retry_exhausted,m10_archival_failure_permanent,m10_archival_retry_policy_invalid,m10_archival_retry_attempt_invalid"
-    ));
-    assert!(DOC.contains(
-        "archival_retry_policy_contract=max_attempts>=1;base_backoff_seconds>=1;max_backoff_seconds>=base_backoff_seconds"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c12_transient_archival_failure_projects_deterministic_retry_window -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c14_archival_retry_budget_exhaustion_and_permanent_failure_fail_closed -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5287"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_phase6_execution_tick_orchestration_markers() {
-    assert!(DOC.contains("## Phase-6 Retention+Archival Execution Tick Contracts (Issue #5289)"));
-    assert!(DOC.contains("phase6_execution_tick_status=verified"));
-    assert!(DOC.contains(
-        "phase6_execution_tick_reason_taxonomy_version=kamn.runtime.data-layer-m10-phase6-execution-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_execution_tick_reason_codes_csv=m10_phase6_execution_applied,m10_phase6_execution_owner_scope_denied,m10_phase6_execution_legal_hold_active,m10_phase6_execution_input_invalid,m10_phase6_execution_projection_input_invalid,m10_phase6_execution_projection_failed"
-    ));
-    assert!(DOC.contains(
-        "phase6_execution_tick_contract=retention_due_lookup->crypto_shred->partition_projection->archive_due"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c16_phase6_orchestration_tick_executes_retention_shred_projection_and_archive -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c19_phase6_orchestration_tick_fails_closed_on_legal_hold_and_empty_projection_entries -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5289"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_phase6_execution_tick_budget_markers() {
-    assert!(DOC.contains("## Phase-6 Execution Tick Budget Guardrail Contracts (Issue #5291)"));
-    assert!(DOC.contains("phase6_execution_tick_budget_status=verified"));
-    assert!(DOC.contains(
-        "phase6_execution_tick_budget_reason_taxonomy_version=kamn.runtime.data-layer-m10-phase6-execution-budget-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_execution_tick_budget_reason_codes_csv=m10_phase6_execution_budget_within_limit,m10_phase6_execution_budget_due_candidates_exceeded,m10_phase6_execution_budget_shredded_messages_exceeded,m10_phase6_execution_budget_projections_exceeded,m10_phase6_execution_budget_archive_entries_exceeded,m10_phase6_execution_budget_invalid"
-    ));
-    assert!(DOC.contains(
-        "phase6_execution_tick_budget_contract=due_candidates<=max_due_candidates;shredded_messages<=max_shredded_messages;projection_reports<=max_projection_reports;archived_entries<=max_archived_entries"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c20_phase6_execution_tick_budget_within_limits_and_exceeded_paths_are_deterministic -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c22_phase6_execution_tick_budget_invalid_limits_fail_closed -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5291"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_phase6_scheduler_cycle_markers() {
-    assert!(DOC.contains(
-        "## Phase-6 Scheduler Cycle Trigger and Guarded Execution Contracts (Issue #5293)"
-    ));
-    assert!(DOC.contains("phase6_scheduler_cycle_status=verified"));
-    assert!(DOC.contains(
-        "phase6_scheduler_trigger_reason_taxonomy_version=kamn.runtime.data-layer-m10-phase6-scheduler-trigger-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_scheduler_trigger_reason_codes_csv=m10_phase6_scheduler_trigger_deferred,m10_phase6_scheduler_trigger_due_threshold,m10_phase6_scheduler_trigger_interval_elapsed"
-    ));
-    assert!(DOC.contains(
-        "phase6_scheduler_cycle_reason_taxonomy_version=kamn.runtime.data-layer-m10-phase6-scheduler-cycle-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_scheduler_cycle_reason_codes_csv=m10_phase6_scheduler_cycle_deferred,m10_phase6_scheduler_cycle_applied,m10_phase6_scheduler_policy_invalid,m10_phase6_scheduler_signal_invalid,m10_phase6_execution_budget_due_candidates_exceeded,m10_phase6_execution_budget_shredded_messages_exceeded,m10_phase6_execution_budget_projections_exceeded,m10_phase6_execution_budget_archive_entries_exceeded"
-    ));
-    assert!(DOC.contains(
-        "phase6_scheduler_cycle_contract=trigger_decision->preflight_budget_admission->phase6_execution_tick->budget_evidence"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c23_phase6_scheduler_trigger_decision_orders_due_threshold_interval_and_deferred -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c25_phase6_scheduler_cycle_preflight_budget_overflow_fails_closed_before_execution -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5293"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_phase6_scheduler_runtime_checkpoint_markers() {
-    assert!(
-        DOC.contains("## Phase-6 Stateful Scheduler Runtime Checkpoint Contracts (Issue #5295)")
-    );
-    assert!(DOC.contains("phase6_scheduler_runtime_checkpoint_status=verified"));
-    assert!(DOC.contains(
-        "phase6_scheduler_runtime_reason_taxonomy_version=kamn.runtime.data-layer-m10-phase6-scheduler-runtime-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_scheduler_runtime_reason_codes_csv=m10_phase6_scheduler_runtime_initialized,m10_phase6_scheduler_cycle_deferred,m10_phase6_scheduler_cycle_applied,m10_phase6_scheduler_signal_invalid,m10_phase6_execution_budget_due_candidates_exceeded"
-    ));
-    assert!(DOC.contains(
-        "phase6_scheduler_runtime_state_contract=total_cycles=executed_cycles+deferred_cycles+fail_closed_cycles;last_successful_tick_epoch_seconds_updates_on_applied_only;last_observed_now_epoch_seconds_monotonic_non_decreasing"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c28_phase6_scheduler_runtime_initializes_zeroed_state_and_checkpoint -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c31_phase6_scheduler_runtime_preflight_fail_closed_increments_fail_counter_without_checkpoint_advance -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5295"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_phase6_runtime_evidence_bundle_markers() {
-    assert!(DOC.contains("## Phase-6 Runtime Evidence Bundle Projection Contracts (Issue #5297)"));
-    assert!(DOC.contains("phase6_runtime_evidence_bundle_status=verified"));
-    assert!(DOC.contains(
-        "phase6_runtime_evidence_reason_taxonomy_version=kamn.runtime.data-layer-m10-phase6-runtime-evidence-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_runtime_evidence_reason_codes_csv=m10_phase6_runtime_evidence_applied,m10_phase6_runtime_evidence_deferred,m10_phase6_runtime_evidence_input_invalid"
-    ));
-    assert!(DOC.contains(
-        "phase6_runtime_evidence_bundle_contract=cycle_report+runtime_state->canonical_evidence_bundle;applied_requires_execution_and_budget_payload;deferred_requires_empty_execution_payload"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c33_phase6_runtime_evidence_bundle_projects_applied_cycle_with_deterministic_artifacts -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_m10_partition_archival spec_c35_phase6_runtime_evidence_bundle_fails_closed_when_applied_payload_is_incomplete -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5297"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_phase6_daemon_runtime_integration_markers() {
-    assert!(DOC.contains("## Phase-6 Daemon Runtime Integration Contracts (Issue #5299)"));
-    assert!(DOC.contains("phase6_daemon_runtime_contract_status=verified"));
-    assert!(DOC.contains(
-        "phase6_daemon_runtime_reason_taxonomy_version=kamn.runtime.daemon.phase6.reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_daemon_runtime_reason_codes_csv=m10_phase6_scheduler_cycle_applied,m10_phase6_scheduler_cycle_deferred,m10_phase6_scheduler_signal_invalid,m10_phase6_execution_budget_due_candidates_exceeded"
-    ));
-    assert!(DOC.contains(
-        "phase6_daemon_runtime_contract=daemon_tick_executes_m10_scheduler_runtime;report_projects_phase6_reason_and_counters;clock_regression_fails_closed"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_projects_phase6_applied_runtime_markers_in_report_output -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::regression_daemon_phase6_runtime_projection_fail_closed_reason_is_stable_on_clock_regression -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5299"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_validation_slice_markers() {
-    assert!(DOC.contains(
-        "## PostgreSQL Live Integration + Daemon Runtime Validation Slice (Issue #5338)"
-    ));
-    assert!(DOC.contains("phase6_live_postgres_daemon_runtime_slice_status=verified"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_slice_env_gate=KAMN_TEST_POSTGRES_URL|DATABASE_URL"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_slice_reason_taxonomy_version=kamn.runtime.daemon.phase6-live-postgres.reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_slice_reason_codes_csv=live_postgres_env_unset,live_postgres_adapter_connected,m10_phase6_scheduler_cycle_applied,m10_phase6_scheduler_cycle_deferred"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_slice_contract=live_postgres_env_gate->adapter_connect_and_migrate->daemon_phase6_runtime_projection"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test data_layer_postgres_execution_adapter spec_c01_and_c03_live_adapter_executes_insert_and_lookup_with_session_context -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5338"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_gate_and_deferred_markers() {
-    assert!(DOC.contains("### Gate and Deferred Path Hardening (Issue #5340)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_gate_reason_contract=env_unset->skip_with_reason;env_set->adapter_connect_and_migrate"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_deferred_contract=live_postgres_adapter_connected+shutdown_signal->m10_phase6_scheduler_cycle_deferred"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::regression_runtime_daemon_live_postgres_validation_slice_reports_unset_env_gate_reason -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::unit_runtime_daemon_live_postgres_validation_slice_prefers_kamn_test_postgres_url -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_deferred_path -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5340"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_stability_markers() {
-    assert!(DOC.contains("### Scenario Matrix Stability (Issue #5342)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_contract=env_unset->live_postgres_env_unset;env_set_no_shutdown->m10_phase6_scheduler_cycle_applied;env_set_shutdown->m10_phase6_scheduler_cycle_deferred"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_stability_contract=repeated_runs_preserve_reason_code_per_matrix_scenario"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_env_matrix_contract_is_deterministic -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_matrix_reasons_are_stable_across_repeated_runs -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5342"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_taxonomy_ordering_markers(
-) {
-    assert!(DOC.contains("### Matrix Taxonomy and Canonical Ordering (Issue #5344)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_reason_taxonomy_version=kamn.runtime.daemon.phase6-live-postgres-matrix.reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_reason_codes_csv=live_postgres_env_unset,m10_phase6_scheduler_cycle_applied,m10_phase6_scheduler_cycle_deferred"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_scenarios_csv=env_unset,env_set_no_shutdown,env_set_shutdown"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_order_contract=matrix_rows_order=env_unset->env_set_no_shutdown->env_set_shutdown;reason_codes_align_with_scenarios"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_matrix_projection_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_taxonomy_ordering_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5344"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_taxonomy_bridge_markers(
-) {
-    assert!(DOC.contains("### Runtime-to-Matrix Taxonomy Bridge (Issue #5346)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_reason_taxonomy_version=kamn.runtime.daemon.phase6.reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_reason_taxonomy_version=kamn.runtime.daemon.phase6-live-postgres-matrix.reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_taxonomy_bridge_contract=runtime_reason_taxonomy_v1->matrix_scenario_taxonomy_v1;applied_and_deferred_reasons_must_align"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_matrix_taxonomy_bridge_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_matrix_taxonomy_versions_are_stable_across_repeated_runs -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_taxonomy_bridge_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5346"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_load_profile_markers()
-{
-    assert!(DOC.contains("### Bounded Load-Profile Matrix (Issue #5348)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_load_profile_ids_csv=applied_t3_i10,applied_t5_i25,applied_t9_i40,deferred_t5_i25_s3_d2_to4,deferred_t7_i25_s3_d2_to4,deferred_t9_i40_s3_d2_to4"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_load_profile_contract=applied_profiles->m10_phase6_scheduler_cycle_applied;deferred_profiles->m10_phase6_scheduler_cycle_deferred;runtime_taxonomy_version_stable"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_load_profile_matrix_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_load_profile_matrix_is_deterministic -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_load_profile_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5348"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_role_profile_markers()
-{
-    assert!(DOC.contains("### Role-Profile Matrix Determinism (Issue #5350)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_role_profile_ids_csv=processor_applied,processor_deferred,listener_applied,listener_deferred,approver_applied,approver_deferred"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_role_profile_contract=processor|listener|approver_applied->m10_phase6_scheduler_cycle_applied;processor|listener|approver_deferred->m10_phase6_scheduler_cycle_deferred;runtime_taxonomy_version_stable"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_role_profile_matrix_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_role_profile_matrix_is_deterministic -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_role_profile_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5350"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_role_pair_markers() {
-    assert!(DOC.contains("### Two-Node Role-Pair Matrix (Issue #5352)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_role_pair_ids_csv=processor_to_listener_applied,processor_to_listener_deferred,listener_to_approver_applied,listener_to_approver_deferred,approver_to_processor_applied,approver_to_processor_deferred"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_role_pair_contract=role_pair_leg_a_applied->m10_phase6_scheduler_cycle_applied;role_pair_leg_b_applied->m10_phase6_scheduler_cycle_applied;role_pair_leg_a_deferred->m10_phase6_scheduler_cycle_deferred;role_pair_leg_b_deferred->m10_phase6_scheduler_cycle_deferred;runtime_taxonomy_version_stable"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_role_pair_matrix_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_role_pair_matrix_is_deterministic -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_role_pair_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5352"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_role_pair_markers(
-) {
-    assert!(DOC.contains("### Bounded Parallel Role-Pair Lane Matrix (Issue #5354)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_parallel_role_pair_lane_ids_csv=processor_listener_parallel_applied,processor_listener_parallel_deferred,listener_approver_parallel_applied,listener_approver_parallel_deferred"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_parallel_role_pair_contract=parallel_lane_leg_a_applied->m10_phase6_scheduler_cycle_applied;parallel_lane_leg_b_applied->m10_phase6_scheduler_cycle_applied;parallel_lane_leg_a_deferred->m10_phase6_scheduler_cycle_deferred;parallel_lane_leg_b_deferred->m10_phase6_scheduler_cycle_deferred;runtime_taxonomy_version_stable"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_role_pair_lane_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_role_pair_lane_is_deterministic -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_role_pair_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5354"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_asymmetric_parallel_lane_markers(
-) {
-    assert!(DOC.contains("### Asymmetric Parallel Lane Matrix (Issue #5356)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_asymmetric_parallel_lane_ids_csv=processor_listener_asymmetric_parallel_applied,processor_listener_asymmetric_parallel_deferred,listener_approver_asymmetric_parallel_applied,listener_approver_asymmetric_parallel_deferred"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_asymmetric_parallel_contract=asymmetric_parallel_leg_a_applied->m10_phase6_scheduler_cycle_applied;asymmetric_parallel_leg_b_applied->m10_phase6_scheduler_cycle_applied;asymmetric_parallel_leg_a_deferred->m10_phase6_scheduler_cycle_deferred;asymmetric_parallel_leg_b_deferred->m10_phase6_scheduler_cycle_deferred;runtime_taxonomy_version_stable"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_asymmetric_parallel_lane_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_asymmetric_parallel_lane_is_deterministic -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_asymmetric_parallel_lane_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5356"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_order_invariance_markers(
-) {
-    assert!(DOC.contains("### Parallel Lane Order-Invariance Matrix (Issue #5358)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_order_invariance_contract=baseline_and_permuted_lane_orders_must_produce_equivalent_sorted_reason_taxonomy_fingerprints"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_order_invariance_lane_sets_csv=symmetric_parallel,asymmetric_parallel"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_order_invariance_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_order_is_invariant -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_order_invariance_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5358"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_permutation_invariance_markers(
-) {
-    assert!(DOC.contains("### Parallel Lane Permutation-Invariance Matrix (Issue #5360)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_permutation_ids_csv=baseline,reverse,rotate_left_1,interleaved_even_then_odd"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_matrix_permutation_invariance_contract=deterministic_permutations_must_preserve_sorted_lane_reason_taxonomy_fingerprints"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_permutation_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_permutations_are_invariant -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_permutation_invariance_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5360"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_fingerprint_schema_markers(
-) {
-    assert!(DOC.contains("### Parallel Lane Fingerprint Schema Contracts (Issue #5362)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_fingerprint_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-fingerprint.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_fingerprint_field_order_csv=lane_id,leg_a_reason,leg_a_taxonomy,leg_b_reason,leg_b_taxonomy"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_fingerprint_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_fingerprint_schema_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_fingerprint_schema_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5362"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_scope_markers(
-) {
-    assert!(DOC.contains("### Parallel Lane Topology-Scope Contracts (Issue #5364)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_ids_csv=same_host_parallel,distributed_label_parallel"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_contract=topology_labels_must_preserve_sorted_lane_reason_taxonomy_fingerprints_under_repeated_runs"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_scope_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_scope_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_scope_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5364"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_permutation_markers(
-) {
-    assert!(
-        DOC.contains("### Parallel Lane Topology Permutation-Invariance Contracts (Issue #5366)")
-    );
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_permutation_ids_csv=baseline,reverse,rotate_left_1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_permutation_contract=deterministic_topology_profile_permutations_must_preserve_sorted_topology_fingerprint_bundles"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_permutation_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_permutations_are_invariant -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_permutation_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5366"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_pair_markers(
-) {
-    assert!(DOC.contains("### Parallel Lane Topology Host-Pair Contracts (Issue #5368)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-pair.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_required_host_pair_ids_csv=node_alpha->node_alpha,node_alpha->node_beta"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_contract=host_pair_ids_must_remain_stable_under_repeated_runs_and_topology_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_pair_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_pairs_are_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_pair_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5368"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_pair_directionality_markers(
-) {
-    assert!(
-        DOC.contains("### Parallel Lane Topology Host-Pair Directionality Contracts (Issue #5370)")
-    );
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_directionality_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-pair-directionality.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_directionality_extraction_rule=host_a_to_host_b_arrow_notation_non_commutative"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_directionality_forbidden_reverse_pairs_csv=node_beta->node_alpha"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_pair_directionality_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_pair_directionality_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_pair_directionality_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5370"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_pair_mapping_markers(
-) {
-    assert!(DOC.contains("### Parallel Lane Topology Host-Pair Mapping Contracts (Issue #5372)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_mapping_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-pair-mapping.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_mapping_rows_csv=same_host_parallel->node_alpha->node_alpha,distributed_label_parallel->node_alpha->node_beta"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_mapping_contract=topology_id_to_host_pair_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_pair_mapping_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_pair_mapping_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_pair_mapping_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5372"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_lane_set_mapping_markers(
-) {
-    assert!(DOC.contains("### Parallel Lane Topology Lane-Set Mapping Contracts (Issue #5374)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_lane_set_mapping_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-lane-set-mapping.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_lane_set_mapping_rows_csv=same_host_parallel->symmetric_parallel,distributed_label_parallel->asymmetric_parallel"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_lane_set_mapping_contract=topology_id_to_lane_set_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_lane_set_mapping_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_lane_set_mapping_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_lane_set_mapping_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5374"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_lane_count_mapping_markers(
-) {
-    assert!(DOC.contains("### Parallel Lane Topology Lane-Count Mapping Contracts (Issue #5376)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_lane_count_mapping_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-lane-count-mapping.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_lane_count_mapping_rows_csv=same_host_parallel->4,distributed_label_parallel->4"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_lane_count_mapping_contract=topology_id_to_lane_count_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_lane_count_mapping_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_lane_count_mapping_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_lane_count_mapping_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5376"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_mapping_markers(
-) {
-    assert!(DOC.contains("### Parallel Lane Topology Host-Mode Mapping Contracts (Issue #5378)"));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_mapping_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-mapping.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_mapping_rows_csv=same_host_parallel->same_host,distributed_label_parallel->distributed_label"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_mapping_contract=topology_id_to_host_mode_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_mapping_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_mapping_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_mapping_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5378"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_cardinality_mapping_markers(
-) {
-    assert!(
-        DOC.contains("### Parallel Lane Topology Host-Cardinality Mapping Contracts (Issue #5380)")
-    );
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_cardinality_mapping_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-cardinality-mapping.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_cardinality_mapping_rows_csv=same_host_parallel->1,distributed_label_parallel->2"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_cardinality_mapping_contract=topology_id_to_unique_host_cardinality_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_cardinality_mapping_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_cardinality_mapping_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_cardinality_mapping_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5380"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_cardinality_coherence_markers(
-) {
-    assert!(DOC.contains(
-        "### Parallel Lane Topology Host-Mode-Cardinality Coherence Contracts (Issue #5382)"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_cardinality_coherence_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-cardinality-coherence.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_cardinality_coherence_rows_csv=same_host_parallel->same_host->1,distributed_label_parallel->distributed_label->2"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_cardinality_coherence_contract=topology_id_to_host_mode_cardinality_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_cardinality_coherence_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_cardinality_coherence_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_cardinality_coherence_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5382"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_pair_cardinality_coherence_markers(
-) {
-    assert!(DOC.contains(
-        "### Parallel Lane Topology Host-Pair-Cardinality Coherence Contracts (Issue #5384)"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_cardinality_coherence_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-pair-cardinality-coherence.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_cardinality_coherence_rows_csv=same_host_parallel->node_alpha->node_alpha->1,distributed_label_parallel->node_alpha->node_beta->2"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_pair_cardinality_coherence_contract=topology_id_to_host_pair_cardinality_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_pair_cardinality_coherence_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_pair_cardinality_coherence_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_pair_cardinality_coherence_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5384"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_coherence_markers(
-) {
-    assert!(DOC.contains(
-        "### Parallel Lane Topology Host-Mode-Host-Pair Coherence Contracts (Issue #5386)"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_coherence_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-host-pair-coherence.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_coherence_rows_csv=same_host_parallel->same_host->node_alpha->node_alpha,distributed_label_parallel->distributed_label->node_alpha->node_beta"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_coherence_contract=topology_id_to_host_mode_host_pair_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_coherence_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_coherence_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_coherence_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5386"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_cardinality_coherence_markers(
-) {
-    assert!(DOC.contains(
-        "### Parallel Lane Topology Host-Mode-Host-Pair-Cardinality Coherence Contracts (Issue #5388)"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_cardinality_coherence_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-host-pair-cardinality-coherence.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_cardinality_coherence_rows_csv=same_host_parallel->same_host->node_alpha->node_alpha->1,distributed_label_parallel->distributed_label->node_alpha->node_beta->2"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_cardinality_coherence_contract=topology_id_to_host_mode_host_pair_cardinality_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_cardinality_coherence_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_cardinality_coherence_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_cardinality_coherence_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5388"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_count_coherence_markers(
-) {
-    assert!(DOC.contains(
-        "### Parallel Lane Topology Host-Mode-Host-Pair-Lane-Set-Lane-Count Coherence Contracts (Issue #5390)"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_count_coherence_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-host-pair-lane-set-lane-count-coherence.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_count_coherence_rows_csv=same_host_parallel->same_host->node_alpha->node_alpha->symmetric_parallel->4,distributed_label_parallel->distributed_label->node_alpha->node_beta->asymmetric_parallel->4"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_count_coherence_contract=topology_id_to_host_mode_host_pair_lane_set_lane_count_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_count_coherence_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_count_coherence_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_count_coherence_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5390"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_id_bundle_coherence_markers(
-) {
-    assert!(DOC.contains(
-        "### Parallel Lane Topology Host-Mode-Host-Pair-Lane-Set-Lane-Id-Bundle Coherence Contracts (Issue #5392)"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_id_bundle_coherence_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-host-pair-lane-set-lane-id-bundle-coherence.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_id_bundle_coherence_rows_csv=same_host_parallel->same_host->node_alpha->node_alpha->symmetric_parallel->listener_approver_parallel_applied+listener_approver_parallel_deferred+processor_listener_parallel_applied+processor_listener_parallel_deferred,distributed_label_parallel->distributed_label->node_alpha->node_beta->asymmetric_parallel->listener_approver_asymmetric_parallel_applied+listener_approver_asymmetric_parallel_deferred+processor_listener_asymmetric_parallel_applied+processor_listener_asymmetric_parallel_deferred"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_id_bundle_coherence_contract=topology_id_to_host_mode_host_pair_lane_set_lane_id_bundle_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_id_bundle_coherence_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_id_bundle_coherence_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_id_bundle_coherence_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5392"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_bundle_coherence_markers(
-) {
-    assert!(DOC.contains(
-        "### Parallel Lane Topology Host-Mode-Host-Pair-Lane-Set-Lane-Fingerprint-Bundle Coherence Contracts (Issue #5394)"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_bundle_coherence_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-host-pair-lane-set-lane-fingerprint-bundle-coherence.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_bundle_coherence_rows_csv=same_host_parallel->same_host->node_alpha->node_alpha->symmetric_parallel->listener_approver_parallel_applied|m10_phase6_scheduler_cycle_applied|kamn.runtime.daemon.phase6.reason-taxonomy.v1|m10_phase6_scheduler_cycle_applied|kamn.runtime.daemon.phase6.reason-taxonomy.v1+listener_approver_parallel_deferred|m10_phase6_scheduler_cycle_deferred|kamn.runtime.daemon.phase6.reason-taxonomy.v1|m10_phase6_scheduler_cycle_deferred|kamn.runtime.daemon.phase6.reason-taxonomy.v1+processor_listener_parallel_applied|m10_phase6_scheduler_cycle_applied|kamn.runtime.daemon.phase6.reason-taxonomy.v1|m10_phase6_scheduler_cycle_applied|kamn.runtime.daemon.phase6.reason-taxonomy.v1+processor_listener_parallel_deferred|m10_phase6_scheduler_cycle_deferred|kamn.runtime.daemon.phase6.reason-taxonomy.v1|m10_phase6_scheduler_cycle_deferred|kamn.runtime.daemon.phase6.reason-taxonomy.v1,distributed_label_parallel->distributed_label->node_alpha->node_beta->asymmetric_parallel->listener_approver_asymmetric_parallel_applied|m10_phase6_scheduler_cycle_applied|kamn.runtime.daemon.phase6.reason-taxonomy.v1|m10_phase6_scheduler_cycle_applied|kamn.runtime.daemon.phase6.reason-taxonomy.v1+listener_approver_asymmetric_parallel_deferred|m10_phase6_scheduler_cycle_deferred|kamn.runtime.daemon.phase6.reason-taxonomy.v1|m10_phase6_scheduler_cycle_deferred|kamn.runtime.daemon.phase6.reason-taxonomy.v1+processor_listener_asymmetric_parallel_applied|m10_phase6_scheduler_cycle_applied|kamn.runtime.daemon.phase6.reason-taxonomy.v1|m10_phase6_scheduler_cycle_applied|kamn.runtime.daemon.phase6.reason-taxonomy.v1+processor_listener_asymmetric_parallel_deferred|m10_phase6_scheduler_cycle_deferred|kamn.runtime.daemon.phase6.reason-taxonomy.v1|m10_phase6_scheduler_cycle_deferred|kamn.runtime.daemon.phase6.reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_bundle_coherence_contract=topology_id_to_host_mode_host_pair_lane_set_lane_fingerprint_bundle_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_bundle_coherence_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_bundle_coherence_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_bundle_coherence_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5394"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_coherence_markers(
-) {
-    assert!(DOC.contains(
-        "### Parallel Lane Topology Host-Mode-Host-Pair-Lane-Set-Lane-Fingerprint-Hash Coherence Contracts (Issue #5396)"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_coherence_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-host-pair-lane-set-lane-fingerprint-hash-coherence.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_coherence_rows_csv=distributed_label_parallel->distributed_label->node_alpha->node_beta->asymmetric_parallel->18ce08940c67c38e,same_host_parallel->same_host->node_alpha->node_alpha->symmetric_parallel->37e351d41d1e30ea"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_coherence_contract=topology_id_to_host_mode_host_pair_lane_set_lane_fingerprint_hash_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_coherence_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_coherence_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_coherence_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5396"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_markers(
-) {
-    assert!(DOC.contains(
-        "### Parallel Lane Topology Host-Mode-Host-Pair-Lane-Set-Lane-Fingerprint-Hash Order-Normalization Contracts (Issue #5398)"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-host-pair-lane-set-lane-fingerprint-hash-order-normalization.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_rows_csv=distributed_label_parallel->distributed_label->node_alpha->node_beta->asymmetric_parallel->18ce08940c67c38e,same_host_parallel->same_host->node_alpha->node_alpha->symmetric_parallel->37e351d41d1e30ea"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_contract=topology_id_to_host_mode_host_pair_lane_set_lane_fingerprint_hash_rows_must_remain_canonically_sorted_after_order_normalization"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5398"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_digest_markers(
-) {
-    assert!(DOC.contains(
-        "### Parallel Lane Topology Host-Mode-Host-Pair-Lane-Set-Lane-Fingerprint-Hash Order-Normalization-Digest Contracts (Issue #5400)"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_digest_schema_version=kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-host-pair-lane-set-lane-fingerprint-hash-order-normalization-digest.v1"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_digest_rows_csv=distributed_label_parallel->distributed_label->node_alpha->node_beta->asymmetric_parallel->18ce08940c67c38e,same_host_parallel->same_host->node_alpha->node_alpha->symmetric_parallel->37e351d41d1e30ea"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_digest_rows_fnv1a64_hex=25b9729eaeb44fe9"
-    ));
-    assert!(DOC.contains(
-        "phase6_live_postgres_daemon_runtime_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_digest_contract=topology_id_to_host_mode_host_pair_lane_set_lane_fingerprint_hash_rows_digest_must_remain_stable_under_order_normalization_and_permutations"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_digest_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_digest_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_live_postgres_daemon_runtime_matrix_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_digest_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5400"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_daemon_tests_live_postgres_fixture_decomposition_markers()
-{
-    assert!(DOC.contains(
-        "### Daemon Tests Live-Postgres Fixture Decomposition Contracts (Issue #5402, #5418, #5420)"
-    ));
-    assert!(DOC.contains(
-        "daemon_tests_live_postgres_fixture_module_path=crates/kamn-node/src/main_tests/daemon_tests/live_postgres_fixtures.rs"
-    ));
-    assert!(DOC.contains("daemon_tests_live_postgres_fixture_phase1_target_max_lines=4300"));
-    assert!(DOC.contains(
-        "daemon_tests_live_postgres_fixture_phase2_module_path=crates/kamn-node/src/main_tests/daemon_tests/live_postgres_topology_contract_tests.rs"
-    ));
-    assert!(DOC.contains("daemon_tests_live_postgres_fixture_phase2_target_max_lines=2200"));
-    assert!(DOC.contains(
-        "daemon_tests_live_postgres_fixture_phase3_runtime_module_path=crates/kamn-node/src/main_tests/daemon_tests/runtime_contract_tests.rs"
-    ));
-    assert!(DOC.contains(
-        "daemon_tests_live_postgres_fixture_phase3_matrix_module_path=crates/kamn-node/src/main_tests/daemon_tests/live_postgres_matrix_contract_tests.rs"
-    ));
-    assert!(DOC.contains("daemon_tests_live_postgres_fixture_phase3_root_target_max_lines=300"));
-    assert!(DOC.contains(
-        "daemon_tests_live_postgres_fixture_test_path_contract=main_tests::daemon_tests::path_prefix_must_remain_stable_after_fixture_extraction"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_digest_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_daemon_tests_live_postgres_fixture_decomposition_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5402, #5418, #5420"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_multi_host_batched_coherence_bundle_markers() {
-    assert!(DOC.contains(
-        "### Multi-Host Daemon Live-Postgres Batched Coherence Bundle Map (Issue #5422)"
-    ));
-    assert!(DOC.contains("daemon_live_postgres_coherence_bundle_b01=runtime_matrix_bundle"));
-    assert!(DOC.contains("daemon_live_postgres_coherence_bundle_b02=parallel_lane_bundle"));
-    assert!(DOC.contains("daemon_live_postgres_coherence_bundle_b03=topology_mapping_bundle"));
-    assert!(DOC.contains("daemon_live_postgres_coherence_bundle_b04=topology_coherence_bundle"));
-    assert!(DOC.contains("daemon_live_postgres_coherence_bundle_b05=fingerprint_stability_bundle"));
-    assert!(DOC.contains("daemon_live_postgres_coherence_bundle_b06=multi_host_execution_bundle"));
-    assert!(DOC.contains("daemon_live_postgres_coherence_bundle_issue_ceiling=8"));
-    assert!(DOC.contains("daemon_live_postgres_coherence_bundle_issue_floor=5"));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_multi_host_batched_coherence_bundle_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5422"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_multi_host_distributed_execution_bundle_markers() {
-    assert!(DOC.contains("### Multi-Host Distributed Execution Lane Contracts (Issue #5422)"));
-    assert!(DOC.contains(
-        "daemon_live_postgres_multi_host_execution_reason_taxonomy_version=kamn.runtime.daemon.phase6-live-postgres.multi-host-execution.reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "daemon_live_postgres_multi_host_execution_prerequisite_env_keys_csv=KAMN_TEST_POSTGRES_URL|DATABASE_URL,KAMN_TEST_LIVE_POSTGRES_DISTRIBUTED_HOSTS"
-    ));
-    assert!(DOC.contains(
-        "daemon_live_postgres_multi_host_execution_prerequisite_reason_codes_csv=live_postgres_multi_host_prerequisites_ready,live_postgres_multi_host_prerequisites_missing,live_postgres_multi_host_host_pair_invalid"
-    ));
-    assert!(DOC.contains(
-        "daemon_live_postgres_multi_host_execution_bundle_selector_prefix=main_tests::daemon_tests::"
-    ));
-    assert!(DOC.contains(
-        "daemon_live_postgres_multi_host_execution_bundle_selector_rows_csv=b01_runtime_matrix_bundle->main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_matrix_reasons_are_stable_across_repeated_runs,b02_parallel_lane_bundle->main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_fingerprint_schema_is_stable,b03_topology_mapping_bundle->main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_scope_is_stable,b04_topology_coherence_bundle->main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_id_bundle_coherence_is_stable,b05_fingerprint_stability_bundle->main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_host_pair_lane_set_lane_fingerprint_hash_order_normalization_digest_is_stable,b06_multi_host_execution_bundle->main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_multi_host_execution_bundle_is_stable"
-    ));
-    assert!(DOC.contains(
-        "daemon_live_postgres_multi_host_execution_projection_digest_fnv1a64_hex=25b9729eaeb44fe9"
-    ));
-    assert!(DOC.contains(
-        "daemon_live_postgres_multi_host_execution_contract=prerequisite_gate->distributed_label_projection->fingerprint_digest_stability"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_multi_host_execution_prerequisite_guard_contract_is_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::functional_runtime_daemon_live_postgres_validation_slice_multi_host_execution_bundle_selector_rows_are_canonical -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::integration_runtime_daemon_phase6_live_postgres_validation_slice_multi_host_execution_bundle_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-core --test service_api_ops_configuration_docs service_api_ops_configuration_contains_multi_host_distributed_execution_bundle_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5422"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_convergence_promotion_marker_contracts() {
-    assert!(DOC.contains("## Convergence Promotion Marker Contracts (Issue #5301)"));
-    assert!(DOC.contains("convergence_promotion_contract_status=verified"));
-    assert!(DOC.contains(
-        "convergence_reason_taxonomy_version=kamn.runtime.daemon.convergence.reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "convergence_reason_codes_csv=convergence_promotion_gate_go,convergence_schema_drift_detected,convergence_error_path_drift_detected,convergence_concurrency_drift_detected,convergence_performance_budget_exceeded,convergence_cost_budget_exceeded"
-    ));
-    assert!(DOC.contains(
-        "convergence_promotion_contract=schema+error_path+concurrency+performance+cost->decision;any_failed_gate=no_go"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::regression_daemon_convergence_projection_fail_closed_reason_is_stable -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --bin kamn-node main_tests::daemon_tests::regression_runtime_daemon_shutdown_timeout_emits_structured_timeout_drain_markers -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #5301"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_retention_policy_fixture_matrix_controls() {
-    assert!(DOC
-        .contains("## Retention Policy Fixture Matrix and Parser Helper Contracts (Issue #4075)"));
-    assert!(DOC.contains(
-        "retention_policy_fixture_matrix_path=fixtures/runtime/retention_policy_fixture_matrix.txt"
-    ));
-    assert!(DOC.contains(
-        "retention_policy_fixture_matrix_schema_version=kamn.runtime.retention-policy-fixture-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "retention_policy_reason_taxonomy_version=kamn.runtime.retention-policy-fixture-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "retention_policy_reason_codes_csv=retention_domain_unknown,retention_window_non_positive"
-    ));
-    assert!(DOC.contains(
-        "retention_policy_fixture_columns=case_id|domain|max_age_seconds|expected_status|expected_reason_code"
-    ));
-    assert!(DOC.contains("retention_domain_unknown"));
-    assert!(DOC.contains("retention_window_non_positive"));
-    assert!(DOC.contains("cargo test -p kamn-core --test retention_policy_fixture_parser_contract"));
-    assert!(DOC.contains("Regression: #4075"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_deletion_proof_artifact_fixture_controls() {
-    assert!(DOC.contains(
-        "## Deletion-Proof Artifact Fixture Set and Checker Behavior Contracts (Issue #4077)"
-    ));
-    assert!(DOC.contains(
-        "deletion_proof_fixture_matrix_path=fixtures/runtime/deletion_proof_artifact_fixture_matrix.txt"
-    ));
-    assert!(DOC.contains(
-        "deletion_proof_fixture_matrix_schema_version=kamn.runtime.deletion-proof-fixture-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "deletion_proof_reason_taxonomy_version=kamn.runtime.deletion-proof-checker-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "deletion_proof_reason_codes_csv=deletion_proof_subject_missing,deletion_proof_tombstone_missing,deletion_proof_status_invalid,deletion_proof_hash_mismatch"
-    ));
-    assert!(DOC.contains(
-        "deletion_proof_fixture_columns=case_id|subject_id|tombstone_hash|expected_hash|proof_status|expected_status|expected_reason_code"
-    ));
-    assert!(DOC.contains("deletion_proof_subject_missing"));
-    assert!(DOC.contains("deletion_proof_tombstone_missing"));
-    assert!(DOC.contains("deletion_proof_status_invalid"));
-    assert!(DOC.contains("deletion_proof_hash_mismatch"));
-    assert!(DOC.contains("cargo test -p kamn-core --test deletion_proof_artifact_checker_contract"));
-    assert!(DOC.contains("Regression: #4077"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_quota_policy_fixture_matrix_controls() {
-    assert!(
-        DOC.contains("## Quota Policy Fixture Matrix and Parser Helper Contracts (Issue #4090)")
-    );
-    assert!(DOC.contains(
-        "quota_policy_fixture_matrix_path=fixtures/runtime/quota_policy_fixture_matrix.txt"
-    ));
-    assert!(DOC.contains(
-        "quota_policy_fixture_matrix_schema_version=kamn.runtime.quota-policy-fixture-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "quota_policy_reason_taxonomy_version=kamn.runtime.quota-policy-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "quota_policy_reason_codes_csv=quota_scope_unknown,quota_window_non_positive,quota_limit_non_positive"
-    ));
-    assert!(DOC.contains(
-        "quota_policy_fixture_columns=case_id|scope|window_seconds|limit|expected_status|expected_reason_code"
-    ));
-    assert!(DOC.contains("quota_scope_unknown"));
-    assert!(DOC.contains("quota_window_non_positive"));
-    assert!(DOC.contains("quota_limit_non_positive"));
-    assert!(DOC.contains("cargo test -p kamn-core --test quota_policy_fixture_parser_contract"));
-    assert!(DOC.contains("Regression: #4090"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_fairness_starvation_fixture_controls() {
-    assert!(DOC.contains("## Fairness Starvation Fixture and Checker Contracts (Issue #4092)"));
-    assert!(DOC.contains(
-        "fairness_fixture_matrix_path=fixtures/runtime/starvation_fairness_fixture_matrix.txt"
-    ));
-    assert!(DOC.contains(
-        "fairness_fixture_matrix_schema_version=kamn.runtime.fairness-fixture-matrix.v1"
-    ));
-    assert!(DOC.contains(
-        "fairness_reason_taxonomy_version=kamn.runtime.fairness-policy-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "fairness_reason_codes_csv=fairness_scope_unknown,fairness_window_non_positive,fairness_max_gap_non_positive,fairness_weighted_share_exceeds_gap"
-    ));
-    assert!(DOC.contains(
-        "fairness_fixture_columns=case_id|scope|window_seconds|active_weighted_share|max_weighted_share_gap|expected_status|expected_reason_code"
-    ));
-    assert!(DOC.contains("fairness_scope_unknown"));
-    assert!(DOC.contains("fairness_window_non_positive"));
-    assert!(DOC.contains("fairness_max_gap_non_positive"));
-    assert!(DOC.contains("fairness_weighted_share_exceeds_gap"));
-    assert!(DOC.contains("cargo test -p kamn-core --test fairness_policy_checker_contract"));
-    assert!(DOC.contains("Regression: #4092"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_overload_docs_parity_remediation_controls() {
-    assert!(DOC.contains("## Daemon OS-Signal Stress Matrix Overload Profiles (Issue #4094)"));
-    assert!(DOC.contains(
-        "daemon_os_signal_stress_matrix_schema_version=kamn.ci.daemon-os-signal-stress-matrix-report.v1"
-    ));
-    assert!(DOC.contains(
-        "overload_docs_parity_reason_taxonomy_version=kamn.ci.daemon-os-signal-stress-matrix-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "overload_docs_parity_reason_codes_csv=runtime_budget_exceeded,matrix_failure_threshold_exceeded,quarantine_registry_missing,quarantine_reference_present_without_followup,matrix_failures_within_threshold,stable_success_with_quarantine_followup,stable_success"
-    ));
-    assert!(DOC.contains("overload_docs_parity_remediation_map_version=v1"));
-    assert!(DOC.contains(
-        "overload_docs_parity_remediation.runtime_budget_exceeded=reduce iterations or increase max-seconds budget after validating reproducer runtime"
-    ));
-    assert!(DOC.contains(
-        "overload_docs_parity_remediation.matrix_failure_threshold_exceeded=triage failing iteration artifacts and rerun reproducer before promotion"
-    ));
-    assert!(DOC.contains(
-        "overload_docs_parity_remediation.quarantine_registry_missing=restore .ci/flaky-tests.txt or pass an explicit --registry-file"
-    ));
-    assert!(DOC.contains(
-        "overload_docs_parity_remediation.quarantine_reference_present_without_followup=add --quarantine-followup-issue #<id> or retire stale quarantine entries"
-    ));
-    assert!(DOC.contains(
-        "overload_docs_parity_remediation.matrix_failures_within_threshold=track flaky rows and keep threshold + waiver evidence attached to release review"
-    ));
-    assert!(DOC.contains(
-        "overload_docs_parity_remediation.stable_success_with_quarantine_followup=keep follow-up issue open until quarantine references are retired"
-    ));
-    assert!(DOC.contains(
-        "overload_docs_parity_remediation.stable_success=no action required; retain report artifact link in release checklist"
-    ));
-    assert!(DOC.contains("Regression: #4097"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_protocol_mismatch_reason_mapping_controls() {
-    assert!(DOC.contains(
-        "## API Protocol Compliance Mismatch Reason Mapping (Issues #4266, #4270, #4271)"
-    ));
-    assert!(DOC.contains("service_api_axum_protocol_mismatch_reason_mapping_status=verified"));
-    assert!(DOC.contains(
-        "service_api_axum_protocol_mismatch_reason_taxonomy_version=kamn.runtime.service-api-axum-protocol-mismatch-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "service_api_axum_protocol_mismatch_reason_codes_csv=service_api_axum_policy_required_field_missing,service_api_axum_policy_marker_missing,service_api_axum_policy_protocol_taxonomy_mismatch,service_api_axum_policy_limit_contract_mismatch,ci_fast_gate_failed,service_api_axum_policy_expected_decision_mismatch,service_api_axum_policy_violation"
-    ));
-    assert!(DOC.contains("service_api_axum_protocol_mismatch_reason_code=none|<reason>"));
-    assert!(DOC.contains("service_api_axum_policy_protocol_taxonomy_mismatch"));
-    assert!(DOC.contains("service_api_axum_policy_limit_contract_mismatch"));
-    assert!(DOC.contains("Regression: #4270"));
-    assert!(DOC.contains("Regression: #4271"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_audit_integrity_tamper_controls() {
-    assert!(DOC.contains("## Audit Integrity Go/No-Go Policy Controls (Issue #4465)"));
-    assert!(DOC.contains(
-        "audit_integrity_reason_taxonomy_version=kamn.release.gonogo-audit-integrity-convergence-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains("gonogo_audit_integrity_reason_taxonomy_version_mismatch"));
-    assert!(DOC.contains("gonogo_audit_integrity_reason_codes_csv_mismatch"));
-    assert!(DOC.contains("audit integrity gate convergence mismatch"));
-    assert!(DOC.contains("Regression: #4465"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_journal_append_checkpoint_integrity_controls() {
-    assert!(DOC
-        .contains("## Journal Append/Checkpoint Integrity Controls (Issues #4236, #4240, #4241)"));
-    assert!(DOC.contains("append_checkpoint_integrity_status=verified"));
-    assert!(DOC.contains(
-        "append_checkpoint_reason_taxonomy_version=kamn.runtime.append-checkpoint-integrity-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "append_checkpoint_reason_codes_csv=wal_append_marker_missing,wal_checkpoint_marker_missing,append_checkpoint_marker_parity_mismatch"
-    ));
-    assert!(DOC.contains("sqlite_crash_recovery_policy_wal_append_status_mismatch"));
-    assert!(DOC.contains("sqlite_crash_recovery_policy_wal_checkpoint_status_mismatch"));
-    assert!(DOC.contains("sqlite_crash_recovery_policy_append_checkpoint_parity_mismatch"));
-    assert!(DOC.contains("Regression: #4240"));
-    assert!(DOC.contains("Regression: #4241"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_in_memory_provider_rejection_controls() {
-    assert!(DOC.contains("## Production-Mode In-Memory Provider Rejection Controls (Issue #4371)"));
-    assert!(DOC.contains("runtime_commit_in_memory_provider_reference_detected"));
-    assert!(DOC.contains("runtime_commit_policy_check_in_memory_provider_reference_detected"));
-    assert!(DOC.contains("InMemoryKolmeRuntimeCommitClient"));
-    assert!(DOC.contains("test_run_local_kamn_live_runtime_integration_contract_lane.sh"));
-    assert!(DOC.contains("Regression: #4371"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_signer_material_validation_and_fallback_prohibition_contracts(
-) {
-    assert!(DOC.contains(
-        "## Signer Material Validation and Fallback Prohibition Contracts (Issues #4167, #4168)"
-    ));
-    assert!(DOC.contains(
-        "signer_config_reason_taxonomy_version=kamn.kolme.local-live-deployment-preflight-signer-config-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "signer_config_reason_codes_csv=signer_secret_missing,signer_secret_invalid_hex,fallback_signer_secret_present_violation,fallback_signer_secret_checkpoint_reason_mismatch,fallback_signer_secret_remediation_missing"
-    ));
-    assert!(DOC.contains("signer_config_reason_codes_value=none|<csv>"));
-    assert!(DOC.contains("signer_secret_missing"));
-    assert!(DOC.contains("signer_secret_invalid_hex"));
-    assert!(DOC.contains("fallback_signer_secret_present_violation"));
-    assert!(DOC.contains("fallback_signer_secret_checkpoint_reason_mismatch"));
-    assert!(DOC.contains("fallback_signer_secret_remediation_missing"));
-    assert!(DOC.contains(
-        "runtime_signer_key_source_policy_reason_codes_csv=production_signer_key_source_env_local_forbidden,fallback_signer_secret_present_violation"
-    ));
-    assert!(DOC.contains(
-        "managed_signer_provenance_reason_codes_csv=managed_signer_backend_response_provenance_missing,managed_signer_backend_response_provenance_malformed,managed_signer_backend_response_provenance_mismatch"
-    ));
-    assert!(DOC.contains("managed_signer_backend_response_provenance_missing"));
-    assert!(DOC.contains("managed_signer_backend_response_provenance_malformed"));
-    assert!(DOC.contains("managed_signer_backend_response_provenance_mismatch"));
-    assert!(DOC.contains("signer secret env is required for selected profile"));
-    assert!(DOC.contains("fallback signer secret env must not be set"));
-    assert!(DOC.contains("remediation: unset KAMN_KOLME_LIVE_SIGNER_PRIVATE_KEY_HEX_FALLBACK"));
-    assert!(DOC.contains("test_run_local_kolme_live_deployment_preflight_lane.sh"));
-    assert!(DOC.contains("test_check_local_kolme_live_deployment_preflight_policy.sh"));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node --test signer_provenance_fallback_policy_contract -- --nocapture"
-    ));
-    assert!(DOC.contains(
-        "check_local_kolme_live_deployment_preflight_policy.py --report-file /tmp/kolme-local-live-deployment-preflight-summary.json --expected-final-decision GO --ci-fast-gate PASS --output-json /tmp/kolme-local-live-deployment-preflight-policy.json"
-    ));
-    assert!(DOC.contains("Regression: #4167"));
-    assert!(DOC.contains("Regression: #4168"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_managed_key_source_adapter_provenance_mapping() {
-    assert!(DOC.contains("managed_key_source_adapter_provenance_status=verified"));
-    assert!(DOC.contains(
-        "managed_key_source_adapter_provenance_fields_csv=profile,key_source,key_reference_env,signer_public_key_hex"
-    ));
-    assert!(DOC.contains(
-        "managed_key_source_adapter_provenance_reason_codes_csv=managed_signer_provenance_marker_profile_mismatch,managed_signer_provenance_marker_key_source_mismatch,managed_signer_provenance_marker_key_reference_env_mismatch,managed_signer_provenance_marker_public_key_missing"
-    ));
-    assert!(DOC.contains("managed_signer_provenance_marker_profile_mismatch"));
-    assert!(DOC.contains("managed_signer_provenance_marker_key_source_mismatch"));
-    assert!(DOC.contains("managed_signer_provenance_marker_key_reference_env_mismatch"));
-    assert!(DOC.contains("managed_signer_provenance_marker_public_key_missing"));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node signer::managed_backend::tests::unit_managed_key_source_adapter_emits_deterministic_provenance_marker -- --exact"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node signer::tests::regression_managed_key_source_provenance_marker_profile_mismatch_fails_closed -- --exact"
-    ));
-    assert!(DOC.contains("Regression: #3955"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_multi_signer_quorum_signature_decision_controls() {
-    assert!(DOC
-        .contains("## Multi-Signer Profile and Quorum Signature-Decision Controls (Issue #4357)"));
-    assert!(DOC.contains(
-        "signature_decision_reason_taxonomy_version=kamn.kolme.local-kamn-live-runtime-signature-decision-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "signature_decision_reason_codes_csv=runtime_signer_profile_missing,runtime_signer_profile_invalid,runtime_signer_previous_profile_missing,runtime_signer_previous_profile_invalid,runtime_signer_failover_profile_unchanged,runtime_signer_profile_changed_without_failover,runtime_signer_rotation_epoch_stale,runtime_signer_rotation_epoch_regressed,runtime_signer_attestation_schema_invalid,runtime_signer_attestation_required_approvals_invalid,runtime_signer_attestation_approved_signers_invalid,runtime_signer_attestation_approved_signers_not_unique,runtime_signer_attestation_quorum_shortfall,runtime_signer_attestation_profile_not_approved,runtime_signer_quorum_linkage_contract_version_invalid,runtime_signer_quorum_linkage_contract_version_mismatch,runtime_signer_quorum_required_approvals_invalid,runtime_signer_quorum_required_approvals_mismatch,runtime_signer_quorum_approved_signers_count_invalid,runtime_signer_quorum_approved_signers_count_mismatch,runtime_signer_quorum_profile_linked_invalid,runtime_signer_quorum_profile_linked_mismatch,runtime_signer_quorum_satisfied_invalid,runtime_signer_quorum_satisfied_mismatch,runtime_signer_quorum_linked_invalid,runtime_signer_quorum_linkage_drift,runtime_signer_quorum_linkage_violation,runtime_signer_failover_attestation_required_approvals_insufficient,runtime_signer_failover_attestation_previous_profile_not_approved"
-    ));
-    assert!(DOC.contains("signature_decision_reason_codes_value=none|<csv>"));
-    assert!(DOC.contains("runtime_signer_attestation_quorum_shortfall"));
-    assert!(DOC.contains("runtime_signer_quorum_linkage_drift"));
-    assert!(DOC.contains("Regression: #4357"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_signer_quorum_profile_matrix_controls() {
-    assert!(DOC.contains("signer_quorum_profile_matrix_fixture_status=verified"));
-    assert!(DOC.contains(
-        "signer_quorum_profile_matrix_case_labels_csv=linked_non_failover_primary,profile_not_approved_non_failover,quorum_shortfall_non_failover,failover_previous_profile_not_approved,linked_failover_dual_approved"
-    ));
-    assert!(DOC.contains(
-        "signer_quorum_profile_matrix_fail_closed_reason_codes_csv=runtime_signer_quorum_linkage_violation,runtime_signer_attestation_quorum_shortfall,runtime_signer_failover_attestation_previous_profile_not_approved"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node signer::signer_policy::tests::unit_signer_quorum_decision_path_matrix -- --exact --nocapture"
-    ));
-    assert!(DOC.contains(
-        "cargo test -p kamn-node main_tests::signer_tests::integration_kolme_live_signer_preflight_quorum_profile_matrix_paths -- --exact --nocapture"
-    ));
-    assert!(DOC.contains("Regression: #3957"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_retry_envelope_exhaustion_reconnect_bound_governance() {
-    assert!(
-        DOC.contains("### Retry Envelope Exhaustion and Reconnect Bound Governance (Issue #4296)")
-    );
-    assert!(DOC.contains(
-        "reason_taxonomy_version=kamn.runtime.local-retry-diagnostics-reason-taxonomy.v2"
-    ));
-    assert!(DOC.contains(
-        "reason_codes_csv=local_retry_readiness_progress_stalled,local_retry_backoff_jitter_parity_bypass_detected,local_retry_envelope_exhaustion_fail_closed_missing,local_retry_reconnect_attempt_bound_drift,local_retry_reconnect_backoff_bound_drift,ci_local_network_budget_boundary_exceeded"
-    ));
-    assert!(DOC.contains("retry_envelope_exhaustion_fail_closed_status=verified"));
-    assert!(DOC.contains("reconnect_attempt_bound_status=verified"));
-    assert!(DOC.contains("reconnect_backoff_bound_status=verified"));
-    assert!(DOC.contains("retry_envelope_max_attempts=3"));
-    assert!(DOC.contains("retry_envelope_max_backoff_seconds=8"));
-    assert!(DOC.contains("local_retry_envelope_exhaustion_fail_closed_missing"));
-    assert!(DOC.contains("local_retry_reconnect_attempt_bound_drift"));
-    assert!(DOC.contains("local_retry_reconnect_backoff_bound_drift"));
-    assert!(DOC.contains("Regression: #4300"));
-    assert!(DOC.contains("Regression: #4301"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_live_node_drift_marker_mismatch_policy_contracts() {
-    assert!(DOC.contains("### Live-Node Drift Marker Mismatch Policy Contracts (Issue #4281)"));
-    assert!(DOC.contains("failover_promotion_gate_status=verified"));
-    assert!(DOC.contains("live_node_drift_parity_status=verified"));
-    assert!(DOC.contains("ci_local_promotion_budget_boundary_status=verified"));
-    assert!(DOC.contains(
-        "failover_readiness_reason_taxonomy_version=kamn.runtime.failover-readiness-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "failover_readiness_reason_codes_csv=failover_readiness_progress_stalled,live_node_drift_marker_parity_mismatch,ci_local_promotion_budget_boundary_exceeded"
-    ));
-    assert!(DOC.contains("failover_sync_drift_policy_status=verified"));
-    assert!(DOC.contains(
-        "bash scripts/runtime/failover_sync_drill_preflight_contract_lane_contract.sh check-policy"
-    ));
-    assert!(DOC.contains("live_node_drift_marker_parity_mismatch"));
-    assert!(DOC.contains("failover_readiness_progress_stalled"));
-    assert!(DOC.contains("ci_local_promotion_budget_boundary_exceeded"));
-    assert!(DOC.contains("failover_sync_drift_policy_required_field_missing:<field>"));
-    assert!(DOC.contains("failover_sync_drift_policy_reason_taxonomy_version_mismatch"));
-    assert!(DOC.contains("failover_sync_drift_policy_reason_codes_csv_mismatch"));
-    assert!(DOC.contains("Regression: #4285"));
-    assert!(DOC.contains("Regression: #4286"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_shutdown_checkpoint_reconciliation_failure_modes() {
-    assert!(DOC.contains("Shutdown signal failure matrix"));
-    assert!(DOC.contains("full_supervisor_stop_graceful_drain_timeout_contract_mismatch"));
-    assert!(DOC.contains(
-        "shutdown_checkpoint_reconciliation_reason_taxonomy_version=kamn.runtime.shutdown-checkpoint-reconciliation-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains("shutdown_checkpoint_reconciliation_timeout_reason_code_mismatch"));
-    assert!(DOC.contains("shutdown_checkpoint_reconciliation_not_signaled_checkpoint_mismatch"));
-    assert!(DOC.contains("runtime_shutdown_invariant_violation:<reason_code>"));
-    assert!(DOC.contains("Regression: #4332"));
-    assert!(DOC.contains("Regression: #4333"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_full_stack_harness_marker_mismatch_controls() {
-    assert!(DOC.contains(
-        "## Full-Stack Harness Marker Completeness and Parity Mismatch Controls (Issue #4195)"
-    ));
-    assert!(DOC.contains("full_io_harness_marker_completeness_status=verified"));
-    assert!(DOC.contains("full_io_harness_marker_parity_status=verified"));
-    assert!(DOC.contains(
-        "full_io_harness_policy_reason_taxonomy_version=kamn.runtime.full-io-scenario-matrix-policy-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "full_io_harness_policy_reason_codes_csv=full_io_scenario_matrix_policy_process_harness_mismatch,full_io_scenario_matrix_policy_api_route_matrix_mismatch,full_io_scenario_matrix_policy_auth_failure_matrix_mismatch,full_io_scenario_matrix_policy_websocket_matrix_mismatch,full_io_scenario_matrix_policy_multinode_propagation_mismatch,full_io_scenario_matrix_policy_dry_run_command_count_mismatch,full_io_scenario_matrix_policy_dry_run_command_status_mismatch"
-    ));
-    assert!(DOC.contains("full_io_scenario_matrix_policy_process_harness_mismatch"));
-    assert!(DOC.contains("full_io_scenario_matrix_policy_dry_run_command_count_mismatch"));
-    assert!(DOC.contains("full_io_scenario_matrix_policy_dry_run_command_status_mismatch"));
-    assert!(DOC.contains("Regression: #4195"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_upgrade_compatibility_marker_matrix_controls() {
-    assert!(DOC.contains("## Upgrade Compatibility Marker Matrix Controls (Issue #4181)"));
-    assert!(DOC.contains(
-        "check_upgrade_compatibility_marker_matrix_policy.py --version-report-file /tmp/kolme-version-report.json --fork-policy-report-file /tmp/kolme-fork-compatibility-policy-report.json --expected-final-decision GO --ci-fast-gate PASS --output-json /tmp/kolme-upgrade-compatibility-marker-matrix-policy-report.json"
-    ));
-    assert!(DOC.contains(
-        "reason_taxonomy_version=kamn.kolme.upgrade-compatibility-marker-matrix-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "reason_codes_csv=version_report_missing,fork_policy_report_missing,version_report_schema_mismatch,version_report_reason_taxonomy_mismatch,version_report_reason_codes_csv_mismatch,version_report_rehearsal_bypass_guard_status_mismatch,version_report_rehearsal_output_normalization_status_mismatch,fork_policy_report_schema_mismatch,fork_policy_report_reason_taxonomy_mismatch,fork_policy_report_reason_codes_csv_mismatch,fork_policy_report_rehearsal_bypass_guard_status_mismatch,fork_policy_report_rehearsal_output_normalization_status_mismatch,expected_final_decision_mismatch,ci_fast_gate_failed"
-    ));
-    assert!(DOC.contains("version_report_schema_mismatch"));
-    assert!(DOC.contains("fork_policy_report_reason_codes_csv_mismatch"));
-    assert!(DOC.contains("fork_policy_report_rehearsal_bypass_guard_status_mismatch"));
-    assert!(DOC.contains("Regression: #4180"));
-    assert!(DOC.contains("Regression: #4181"));
-}
-
-#[test]
-fn service_api_ops_configuration_contains_partition_healing_mismatch_mapping_controls() {
-    assert!(DOC.contains(
-        "### Block Reconciliation Partition-Healing Mismatch Mapping Contracts (Issues #4251, #4255, #4256)"
-    ));
-    assert!(DOC.contains("partition_healing_mismatch_reason_mapping_status=verified"));
-    assert!(DOC.contains(
-        "partition_healing_mismatch_reason_taxonomy_version=kamn.runtime.block-reconciliation-partition-healing-mismatch-reason-taxonomy.v1"
-    ));
-    assert!(DOC.contains(
-        "partition_healing_mismatch_reason_codes_csv=block_reconciliation_partition_rejoin_policy_required_field_missing,block_reconciliation_partition_rejoin_policy_marker_mismatch,block_reconciliation_partition_rejoin_policy_transport_contract_mismatch,block_reconciliation_partition_rejoin_policy_reconciliation_taxonomy_mismatch,block_reconciliation_partition_rejoin_policy_recovery_contract_mismatch,block_reconciliation_partition_rejoin_policy_reconciliation_reason_codes_invalid,block_reconciliation_partition_rejoin_policy_lane_mode_contract_mismatch,block_reconciliation_partition_rejoin_policy_ci_fast_gate_failed,block_reconciliation_partition_rejoin_policy_expected_decision_mismatch,block_reconciliation_partition_rejoin_policy_violation"
-    ));
-    assert!(DOC.contains("partition_healing_mismatch_reason_code=none|<reason>"));
-    assert!(
-        DOC.contains("block_reconciliation_partition_rejoin_policy_required_field_missing:<field>")
-    );
-    assert!(DOC.contains(
-        "block_reconciliation_partition_rejoin_policy_reconciliation_reason_codes_invalid"
-    ));
-    assert!(DOC.contains("Regression: #4255"));
-    assert!(DOC.contains("Regression: #4256"));
-}
+#[path = "service_api_ops_configuration_docs/dependency_supply_chain_contract_tests.rs"]
+mod dependency_supply_chain_contract_tests;
+#[path = "service_api_ops_configuration_docs/compatibility_resilience_contract_tests.rs"]
+mod compatibility_resilience_contract_tests;
+#[path = "service_api_ops_configuration_docs/phase6_runtime_contract_tests.rs"]
+mod phase6_runtime_contract_tests;
+#[path = "service_api_ops_configuration_docs/live_postgres_matrix_contract_tests.rs"]
+mod live_postgres_matrix_contract_tests;
+#[path = "service_api_ops_configuration_docs/guardrail_signer_contract_tests.rs"]
+mod guardrail_signer_contract_tests;
+#[path = "service_api_ops_configuration_docs/reconciliation_upgrade_contract_tests.rs"]
+mod reconciliation_upgrade_contract_tests;
