+++ crates/kamn-core/tests/service_api_ops_configuration_docs/compatibility_resilience_contract_tests/resilience_local_heavy_contract_tests.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+#[path = "resilience_local_heavy_contract_tests/sqlite_local_heavy_contract_tests.rs"]
+mod sqlite_local_heavy_contract_tests;
+#[path = "resilience_local_heavy_contract_tests/capacity_redaction_contract_tests.rs"]
+mod capacity_redaction_contract_tests;
