+++ crates/kamn-core/tests/service_api_ops_configuration_docs/compatibility_resilience_contract_tests/resilience_local_heavy_contract_tests/capacity_redaction_contract_tests.rs
@@ -1,0 +1,14 @@
+use super::*;
+
+#[test]
+fn service_api_ops_configuration_contains_local_heavy_capacity_load_lane_markers() {
+    assert_doc_contains_all(&["## Local-Heavy Capacity/Load Lane Artifact Contract (Issue #4004)", "local_heavy_capacity_load_lane_schema_version=kamn.runtime.local-heavy-capacity-load-lane-report.v1", "local_heavy_capacity_load_artifact_schema_version=kamn.runtime.local-heavy-capacity-load-artifact-schema.v1", "local_heavy_capacity_load_reason_taxonomy_version=kamn.runtime.local-heavy-capacity-load-reason-taxonomy.v1", "local_heavy_capacity_load_reason_codes_csv=local_heavy_capacity_load_profile_threshold_breach,local_heavy_capacity_load_runtime_budget_exceeded", "local_heavy_capacity_load_required_profiles_csv=baseline,fault", "profile=baseline -> status=pass + final_decision=GO + profile_status=verified", "profile=fault -> status=fail + final_decision=NO-GO + profile_status=failed", "bash scripts/runtime/run_local_heavy_capacity_load_lane.sh --profile baseline --mode dry-run --max-seconds 120 --output-json /tmp/local-heavy-capacity-load-baseline.json", "bash scripts/runtime/run_local_heavy_capacity_load_lane.sh --profile fault --mode dry-run --max-seconds 120 --output-json /tmp/local-heavy-capacity-load-fault.json", "cargo test -p kamn-core --test local_heavy_capacity_load_lane_contract -- --nocapture", "Regression: #4004"]);
+}
+#[test]
+fn service_api_ops_configuration_contains_local_heavy_redaction_validation_lane_markers() {
+    assert_doc_contains_all(&["## Local-Heavy Redaction Validation Lane Artifact Contract (Issue #4079)", "local_heavy_redaction_validation_lane_schema_version=kamn.runtime.local-heavy-redaction-validation-lane-report.v1", "local_heavy_redaction_validation_artifact_schema_version=kamn.runtime.local-heavy-redaction-validation-artifact-schema.v1", "local_heavy_redaction_validation_reason_taxonomy_version=kamn.runtime.local-heavy-redaction-validation-reason-taxonomy.v1", "local_heavy_redaction_validation_reason_codes_csv=local_heavy_redaction_sensitive_pattern_detected,local_heavy_redaction_runtime_budget_exceeded", "local_heavy_redaction_validation_required_profiles_csv=baseline,injected-leak", "profile=baseline -> status=pass + final_decision=GO + profile_status=verified + leak_marker_status=clear", "profile=injected-leak -> status=fail + final_decision=NO-GO + profile_status=failed + leak_marker_status=detected", "profile=baseline -> leak_detection_count=0 + leaked_pattern_ids_csv=none", "profile=injected-leak -> leak_detection_count=3 + leaked_pattern_ids_csv=raw_signer_secret,pii_email,pii_phone", "bash scripts/runtime/run_local_heavy_redaction_validation_lane.sh --profile baseline --mode dry-run --max-seconds 120 --output-json /tmp/local-heavy-redaction-validation-baseline.json", "bash scripts/runtime/run_local_heavy_redaction_validation_lane.sh --profile injected-leak --mode dry-run --max-seconds 120 --output-json /tmp/local-heavy-redaction-validation-injected-leak.json", "cargo test -p kamn-core --test local_heavy_redaction_validation_lane_contract -- --nocapture", "Regression: #4079"]);
+}
+#[test]
+fn service_api_ops_configuration_contains_dependency_local_heavy_deep_scan_lane_markers() {
+    assert_doc_contains_all(&["## Dependency Local-Heavy Deep Scan Lane Artifact Contract (Issue #4032)", "dependency_local_heavy_deep_scan_lane_schema_version=kamn.runtime.dependency-local-heavy-deep-scan-lane-report.v1", "dependency_local_heavy_deep_scan_artifact_schema_version=kamn.runtime.dependency-local-heavy-deep-scan-artifact-schema.v1", "dependency_local_heavy_deep_scan_fixture_schema_version=kamn.ci.dependency-local-heavy-deep-scan-fixture-matrix.v1", "dependency_local_heavy_deep_scan_reason_taxonomy_version=kamn.runtime.dependency-local-heavy-deep-scan-reason-taxonomy.v1", "dependency_local_heavy_deep_scan_reason_codes_csv=dependency_local_heavy_deep_scan_profile_threshold_exceeded,dependency_local_heavy_deep_scan_runtime_budget_exceeded", "dependency_local_heavy_deep_scan_required_profiles_csv=baseline,injected-risk", "profile=baseline -> status=pass + final_decision=GO + profile_status=verified + critical_count=0 + high_count=0", "profile=injected-risk -> status=fail + final_decision=NO-GO + profile_status=failed + critical_count=1 + high_count=2", "bash scripts/runtime/run_dependency_local_heavy_deep_scan_lane.sh --profile baseline --mode dry-run --max-seconds 180 --output-json /tmp/dependency-local-heavy-deep-scan-baseline.json", "bash scripts/runtime/run_dependency_local_heavy_deep_scan_lane.sh --profile injected-risk --mode dry-run --max-seconds 180 --output-json /tmp/dependency-local-heavy-deep-scan-injected-risk.json", "cargo test -p kamn-core --test dependency_local_heavy_deep_scan_lane_contract -- --nocapture", "Regression: #4032"]);
+}
