+++ crates/kamn-core/tests/service_api_ops_configuration_docs/dependency_supply_chain_contract_tests.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+#[path = "dependency_supply_chain_contract_tests/signer_dependency_contract_tests.rs"]
+mod signer_dependency_contract_tests;
+#[path = "dependency_supply_chain_contract_tests/sbom_contract_tests.rs"]
+mod sbom_contract_tests;
