+++ crates/kamn-core/tests/service_api_ops_configuration_docs/guardrail_signer_contract_tests.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+#[path = "guardrail_signer_contract_tests/retention_fairness_contract_tests.rs"]
+mod retention_fairness_contract_tests;
+#[path = "guardrail_signer_contract_tests/signer_quorum_contract_tests.rs"]
+mod signer_quorum_contract_tests;
+#[path = "guardrail_signer_contract_tests/retry_envelope_contract_tests.rs"]
+mod retry_envelope_contract_tests;
