+++ crates/kamn-core/tests/service_api_ops_configuration_docs/guardrail_signer_contract_tests/retry_envelope_contract_tests.rs
@@ -1,0 +1,24 @@
+use super::*;
+
+#[test]
+fn service_api_ops_configuration_contains_retry_envelope_exhaustion_reconnect_bound_governance() {
+    assert!(
+        DOC.contains("### Retry Envelope Exhaustion and Reconnect Bound Governance (Issue #4296)")
+    );
+    assert!(DOC.contains(
+        "reason_taxonomy_version=kamn.runtime.local-retry-diagnostics-reason-taxonomy.v2"
+    ));
+    assert!(DOC.contains(
+        "reason_codes_csv=local_retry_readiness_progress_stalled,local_retry_backoff_jitter_parity_bypass_detected,local_retry_envelope_exhaustion_fail_closed_missing,local_retry_reconnect_attempt_bound_drift,local_retry_reconnect_backoff_bound_drift,ci_local_network_budget_boundary_exceeded"
+    ));
+    assert!(DOC.contains("retry_envelope_exhaustion_fail_closed_status=verified"));
+    assert!(DOC.contains("reconnect_attempt_bound_status=verified"));
+    assert!(DOC.contains("reconnect_backoff_bound_status=verified"));
+    assert!(DOC.contains("retry_envelope_max_attempts=3"));
+    assert!(DOC.contains("retry_envelope_max_backoff_seconds=8"));
+    assert!(DOC.contains("local_retry_envelope_exhaustion_fail_closed_missing"));
+    assert!(DOC.contains("local_retry_reconnect_attempt_bound_drift"));
+    assert!(DOC.contains("local_retry_reconnect_backoff_bound_drift"));
+    assert!(DOC.contains("Regression: #4300"));
+    assert!(DOC.contains("Regression: #4301"));
+}
