+++ crates/kamn-core/tests/service_api_ops_configuration_docs/live_postgres_matrix_contract_tests/topology_coherence_contract_tests.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+#[path = "topology_coherence_contract_tests/host_coherence_contract_tests.rs"]
+mod host_coherence_contract_tests;
+#[path = "topology_coherence_contract_tests/lane_bundle_coherence_contract_tests.rs"]
+mod lane_bundle_coherence_contract_tests;
+#[path = "topology_coherence_contract_tests/digest_normalization_contract_tests.rs"]
+mod digest_normalization_contract_tests;
