+++ crates/kamn-core/tests/service_api_ops_configuration_docs/live_postgres_matrix_contract_tests/topology_scope_contract_tests.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+#[path = "topology_scope_contract_tests/runtime_matrix_scope_contract_tests.rs"]
+mod runtime_matrix_scope_contract_tests;
+#[path = "topology_scope_contract_tests/topology_scope_bundle_contract_tests.rs"]
+mod topology_scope_bundle_contract_tests;
